refactor(settings): reorganize into an index + drill-in structure
Replace the 3,339-line SettingsScreen monolith and its 7-tab bar with a
scrollable Settings index of category rows, each drilling into its own
screen. Splits the overloaded "General" tab into Appearance and Home &
Layout, and gives the settings surface room to grow: adding a feature is
one new SettingsCategory entry plus a screen file.

- New SettingsIndexScreen (landing list) + SettingsCategory model.
- Per-category screens: Appearance, HomeLayout, Games, Media,
  RetroAchievements, SaveSync, Friends, LockedMode. Section composables
  moved verbatim to minimize risk.
- SettingsCommon extracts the shared SettingsDetailScaffold (ThemedScreen
  + top bar + d-pad/A-button handling) and the shared primitives
  (dpadFocusable, SettingsCard, CardSwitchRow, gradientBrush, etc.).
- Settings is now a nested nav graph; all category screens share ONE
  SettingsViewModel scoped to the graph back-stack entry, so credentials
  held in the VM's in-memory state survive the first-launch finish.
- MainActivity settings-active detection matches "settings_*" routes.

Verified on-device (Retroid Pocket 4 Pro, FullDebug): index, drill-in,
back, Games wiring, and a shared-VM toggle round-trip. Full + Lite and
unit/android tests compile.
