+++ app/src/main/java/com/gamelaunch/frontend/ui/navigation/Screen.kt
@@ -4,7 +4,22 @@ sealed class Screen(val route: String) {
     object Onboarding : Screen("onboarding")
     object Home : Screen("home")
     object Scan : Screen("scan")
-    object Settings : Screen("settings")
+
+    /** Nested graph wrapping the settings index + its drill-in category screens. Navigating here
+     * enters [SettingsIndex] (the graph's start). The graph scopes a single shared SettingsViewModel
+     * so credentials typed on any category screen survive the first-launch "Library" finish. */
+    object Settings : Screen("settings_graph")
+    object SettingsIndex : Screen("settings_index")
+    // Category routes — kept in sync with SettingsCategory.route in the settings package.
+    object SettingsAppearance : Screen("settings_appearance")
+    object SettingsHomeLayout : Screen("settings_home_layout")
+    object SettingsGames : Screen("settings_games")
+    object SettingsMedia : Screen("settings_media")
+    object SettingsRetroAchievements : Screen("settings_retro_achievements")
+    object SettingsSaveSync : Screen("settings_save_sync")
+    object SettingsFriends : Screen("settings_friends")
+    object SettingsLocked : Screen("settings_locked")
+
     object LockedModeGames : Screen("locked_mode_games")
     object LockedModeApps : Screen("locked_mode_apps")
     object EmulatorConfig : Screen("emulator_config")
