--- app/src/main/java/com/gamelaunch/frontend/ui/screen/settings/SettingsScreen.kt
@@ -1,3339 +1,0 @@
-package com.gamelaunch.frontend.ui.screen.settings
-
-import android.content.Context
-import android.content.Intent
-import android.Manifest
-import android.content.pm.PackageManager
-import android.os.Build
-import android.hardware.display.DisplayManager
-import android.provider.Settings
-import android.view.Display
-import com.gamelaunch.frontend.BuildConfig
-import androidx.activity.compose.rememberLauncherForActivityResult
-import androidx.activity.result.PickVisualMediaRequest
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.focusGroup
-import androidx.compose.foundation.focusable
-import androidx.compose.foundation.horizontalScroll
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material.icons.automirrored.filled.ArrowForward
-import androidx.compose.material.icons.filled.Check
-import androidx.compose.material.icons.filled.CheckCircle
-import androidx.compose.material.icons.filled.Close
-import androidx.compose.material.icons.filled.RadioButtonUnchecked
-import androidx.compose.material.icons.filled.Sync
-import androidx.compose.material.icons.filled.Group
-import androidx.compose.material.icons.filled.Home
-import androidx.compose.material.icons.filled.Lock
-import androidx.compose.material.icons.filled.Share
-import androidx.compose.material.icons.filled.EmojiEvents
-import com.gamelaunch.frontend.domain.friends.Friend
-import com.gamelaunch.frontend.domain.friends.FriendStatus
-import com.gamelaunch.frontend.systemui.SystemNavigationLockStatus
-import com.gamelaunch.frontend.systemui.SystemNavigationSetupProgress
-import com.gamelaunch.frontend.ui.screen.friends.FriendsViewModel
-import androidx.compose.material.icons.filled.FolderOpen
-import androidx.compose.material.icons.filled.PermMedia
-import androidx.compose.material.icons.filled.Tune
-import androidx.compose.material.icons.filled.VideogameAsset
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.AlertDialog
-import com.gamelaunch.frontend.domain.model.EmulatorUpdate
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.HorizontalDivider
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.LinearProgressIndicator
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.OutlinedTextFieldDefaults
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Slider
-import androidx.compose.material3.SliderDefaults
-import androidx.compose.material3.Snackbar
-import androidx.compose.material3.SnackbarHost
-import androidx.compose.material3.SnackbarHostState
-import androidx.compose.material3.Switch
-import androidx.compose.material3.SwitchDefaults
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
-import androidx.compose.material3.TopAppBar
-import androidx.compose.material3.TopAppBarDefaults
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.DisposableEffect
-import androidx.compose.runtime.MutableState
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.compositionLocalOf
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.key
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.focus.FocusDirection
-import androidx.compose.ui.focus.FocusRequester
-import androidx.compose.ui.focus.focusRequester
-import androidx.compose.ui.input.key.Key
-import androidx.compose.ui.input.key.KeyEventType
-import androidx.compose.ui.input.key.key
-import androidx.compose.ui.input.key.onKeyEvent
-import androidx.compose.ui.input.key.type
-import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.toArgb
-import androidx.compose.ui.graphics.Shape
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.platform.LocalFocusManager
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.ui.platform.LocalConfiguration
-import androidx.compose.ui.platform.LocalContext
-import androidx.core.content.ContextCompat
-import androidx.compose.ui.platform.LocalLifecycleOwner
-import androidx.compose.ui.text.SpanStyle
-import androidx.compose.ui.text.buildAnnotatedString
-import androidx.compose.ui.text.withStyle
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.window.Dialog
-import androidx.compose.ui.window.DialogProperties
-import androidx.compose.material3.Checkbox
-import androidx.compose.material3.CheckboxDefaults
-import androidx.compose.material3.Surface
-import androidx.compose.ui.focus.onFocusChanged
-import com.gamelaunch.frontend.ui.component.AppIcon
-import androidx.compose.ui.text.input.PasswordVisualTransformation
-import androidx.compose.ui.unit.dp
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.LifecycleEventObserver
-import androidx.lifecycle.compose.LifecycleResumeEffect
-import com.gamelaunch.frontend.launcher.HomeLauncherHelper
-import com.gamelaunch.frontend.domain.sync.EmulatorSyncStatus
-import com.gamelaunch.frontend.domain.sync.SyncReadiness
-import com.gamelaunch.frontend.ui.component.QrCode
-import com.gamelaunch.frontend.domain.usecase.EsdeImportStatus
-import com.gamelaunch.frontend.domain.usecase.LbSyncStatus
-import com.gamelaunch.frontend.ui.input.GamepadA
-import com.gamelaunch.frontend.ui.input.GamepadL1
-import com.gamelaunch.frontend.ui.input.GamepadR1
-import com.gamelaunch.frontend.domain.lockedmode.LockedModeState
-import com.gamelaunch.frontend.ui.lockedmode.LockedModeDialogStep
-import com.gamelaunch.frontend.ui.lockedmode.LockedModeActivationDialog
-import com.gamelaunch.frontend.ui.lockedmode.LockedModeSettingsViewModel
-import com.gamelaunch.frontend.ui.lockedmode.PinPadDialog
-import com.gamelaunch.frontend.ui.theme.CardColorScheme
-import com.gamelaunch.frontend.ui.theme.ElectricBlue
-import com.gamelaunch.frontend.ui.theme.LayoutMode
-import com.gamelaunch.frontend.ui.theme.MonochromeSeeds
-import com.gamelaunch.frontend.ui.theme.NeonPurple
-import com.gamelaunch.frontend.ui.theme.ThemedScreen
-import com.gamelaunch.frontend.util.StorageUtils
-import com.journeyapps.barcodescanner.ScanContract
-import com.journeyapps.barcodescanner.ScanOptions
-import kotlinx.coroutines.launch
-import kotlin.math.roundToInt
-
-private val gradientBrush = Brush.horizontalGradient(listOf(ElectricBlue, NeonPurple))
-
-/**
- * Holds the click action of whichever settings control currently has d-pad focus. Compose's
- * [clickable] only self-activates on Enter / DPAD-center, not the gamepad **A** button
- * (KEYCODE_BUTTON_A), so the screen-level key handler reads this to fire A on the focused control.
- */
-internal val LocalFocusedAction = compositionLocalOf<MutableState<(() -> Unit)?>> {
-    error("LocalFocusedAction not provided")
-}
-
-/**
- * Turns a settings control into a d-pad focus target: draws a highlight ring while focused, registers
- * its [onClick] so the screen's A-button handler can activate it, and keeps the ordinary touch click.
- * Use this in place of [clickable] on interactive settings rows/chips/buttons.
- */
-@Composable
-private fun Modifier.dpadFocusable(
-    enabled: Boolean = true,
-    shape: Shape = RoundedCornerShape(10.dp),
-    onClick: () -> Unit
-): Modifier {
-    if (!enabled) return this
-    val focusedAction = LocalFocusedAction.current
-    var focused by remember { mutableStateOf(false) }
-    return this
-        .onFocusChanged {
-            focused = it.isFocused
-            if (it.isFocused) focusedAction.value = onClick
-        }
-        .border(
-            width = if (focused) 2.dp else 0.dp,
-            color = if (focused) ElectricBlue else Color.Transparent,
-            shape = shape
-        )
-        .clickable(onClick = onClick)
-}
-
-private enum class SettingsTab(val label: String, val icon: ImageVector) {
-    GENERAL("General", Icons.Default.Tune),
-    GAMES("Games", Icons.Default.VideogameAsset),
-    MEDIA("Media", Icons.Default.PermMedia),
-    LOCKED_MODE("Locked Mode", Icons.Default.Lock),
-    RETRO_ACHIEVEMENTS("RetroAchievements", Icons.Default.EmojiEvents),
-    SAVE_SYNC("Save Sync", Icons.Default.Sync),
-    FRIENDS("Friends", Icons.Default.Group)
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun SettingsScreen(
-    onBack: (() -> Unit)?,
-    onGoToLibrary: () -> Unit,
-    onEmulatorConfigClick: () -> Unit,
-    onManageAllowedGames: () -> Unit,
-    onManageAllowedApps: () -> Unit,
-    onScrapeAllClick: () -> Unit,
-    onRescanClick: () -> Unit,
-    viewModel: SettingsViewModel = hiltViewModel()
-) {
-    val context = LocalContext.current
-    val state by viewModel.uiState.collectAsState()
-    val snackbarHostState = remember { SnackbarHostState() }
-    val coroutineScope = rememberCoroutineScope()
-    val storageVolumes = remember { StorageUtils.getStorageVolumes(context) }
-
-    var isDefaultHome by remember { mutableStateOf(HomeLauncherHelper.isDefaultHome(context)) }
-    // No-op result callback: the picker/settings screens don't return a reliable result, so the
-    // ON_RESUME observer below refreshes the status when we come back instead.
-    val homeLauncherPicker = rememberLauncherForActivityResult(
-        contract = ActivityResultContracts.StartActivityForResult()
-    ) {}
-
-    fun openHomeLauncherSettings() {
-        val opened = runCatching {
-            homeLauncherPicker.launch(HomeLauncherHelper.selectionIntent(context))
-        }.isSuccess || runCatching {
-            context.startActivity(Intent(Settings.ACTION_SETTINGS))
-        }.isSuccess
-
-        if (!opened) {
-            coroutineScope.launch {
-                snackbarHostState.showSnackbar("Home app settings aren't available on this device")
-            }
-        }
-    }
-
-    // Settings and OEM chooser activities do not consistently return a useful result. Re-checking
-    // on resume makes the status correct regardless of which system UI handled the selection.
-    val lifecycleOwner = LocalLifecycleOwner.current
-    DisposableEffect(lifecycleOwner, context) {
-        val observer = LifecycleEventObserver { _, event ->
-            if (event == Lifecycle.Event.ON_RESUME) {
-                isDefaultHome = HomeLauncherHelper.isDefaultHome(context)
-            }
-        }
-        lifecycleOwner.lifecycle.addObserver(observer)
-        onDispose { lifecycleOwner.lifecycle.removeObserver(observer) }
-    }
-
-    var selectedTab by rememberSaveable { mutableStateOf(SettingsTab.GENERAL) }
-
-    LaunchedEffect(state.emulatorDetectResult) {
-        state.emulatorDetectResult?.let { msg ->
-            snackbarHostState.showSnackbar(msg)
-            viewModel.clearEmulatorDetectResult()
-        }
-    }
-
-    val folderPicker = rememberLauncherForActivityResult(
-        contract = ActivityResultContracts.OpenDocumentTree()
-    ) { uri ->
-        uri?.let {
-            val path = StorageUtils.resolveTreeUriToPath(it) ?: it.toString()
-            viewModel.setRomRootPath(path)
-        }
-    }
-
-    val mediaStoragePicker = rememberLauncherForActivityResult(
-        contract = ActivityResultContracts.OpenDocumentTree()
-    ) { uri ->
-        uri?.let {
-            val path = StorageUtils.resolveTreeUriToPath(it) ?: it.toString()
-            // Persist the folder and auto-import any existing media it contains.
-            viewModel.chooseMediaStorageFolder(path)
-        }
-    }
-
-    // Android photo picker (no storage permission needed) for the custom branded background.
-    val backgroundImagePicker = rememberLauncherForActivityResult(
-        contract = ActivityResultContracts.PickVisualMedia()
-    ) { uri ->
-        uri?.let { viewModel.importBackgroundImage(it) }
-    }
-
-    // The Friends settings tab is always available — turning the feature off only stops sharing and
-    // hides the Home tab, so users can always come back here to turn it on again.
-    val visibleTabs = SettingsTab.entries
-
-    // L1 / R1 cycle between tabs (with wraparound), mirroring the home screen.
-    fun cycleTab(delta: Int) {
-        val cur = visibleTabs.indexOf(selectedTab).coerceAtLeast(0)
-        selectedTab = visibleTabs[(cur + delta + visibleTabs.size) % visibleTabs.size]
-    }
-
-    // D-pad navigation: focus is driven into the scrolling content (see the Column below), and this
-    // screen-level handler turns the d-pad into focus movement plus A-to-activate. focusedAction is
-    // the click of whatever control is focused, so A fires it (Compose's clickable ignores BUTTON_A).
-    val focusManager = LocalFocusManager.current
-    val contentFocus = remember { FocusRequester() }
-    val focusedAction = remember { mutableStateOf<(() -> Unit)?>(null) }
-
-    // Clear any carried-over activation action when the tab changes. We intentionally do NOT
-    // auto-request focus into the content on entry/tab-switch: when the first control is a text
-    // field that pops the soft keyboard on handhelds and makes the screen un-navigable. The d-pad
-    // still grabs focus via moveFocus on the first directional press.
-    LaunchedEffect(selectedTab) {
-        focusedAction.value = null
-    }
-
-    // Honour the user's light/dark choice for this screen's Material components.
-    ThemedScreen {
-        CompositionLocalProvider(LocalFocusedAction provides focusedAction) {
-            Box(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .onKeyEvent { event ->
-                        if (event.type != KeyEventType.KeyDown) return@onKeyEvent false
-                        when (event.key) {
-                            GamepadL1 -> {
-                                cycleTab(-1); true
-                            }
-
-                            GamepadR1 -> {
-                                cycleTab(+1); true
-                            }
-
-                            Key.DirectionDown -> focusManager.moveFocus(FocusDirection.Down)
-                            Key.DirectionUp -> focusManager.moveFocus(FocusDirection.Up)
-                            Key.DirectionLeft -> focusManager.moveFocus(FocusDirection.Left)
-                            Key.DirectionRight -> focusManager.moveFocus(FocusDirection.Right)
-                            GamepadA, Key.DirectionCenter, Key.Enter -> {
-                                val action = focusedAction.value
-                                if (action != null) {
-                                    action(); true
-                                } else false
-                            }
-
-                            else -> false
-                        }
-                    }
-            ) {
-                Scaffold(
-                    containerColor = MaterialTheme.colorScheme.background,
-                    snackbarHost = {
-                        SnackbarHost(snackbarHostState) { data -> Snackbar(snackbarData = data) }
-                    },
-                    topBar = {
-                        TopAppBar(
-                            title = {
-                                Text(
-                                    "Settings",
-                                    style = MaterialTheme.typography.titleLarge,
-                                    modifier = Modifier.padding(start = 4.dp)
-                                )
-                            },
-                            navigationIcon = {
-                                if (onBack != null) {
-                                    IconButton(
-                                        onClick = onBack,
-                                        modifier = Modifier
-                                            .padding(8.dp)
-                                            .size(36.dp)
-                                            .background(
-                                                MaterialTheme.colorScheme.surfaceVariant,
-                                                CircleShape
-                                            )
-                                    ) {
-                                        Icon(
-                                            Icons.AutoMirrored.Filled.ArrowBack,
-                                            contentDescription = "Back",
-                                            modifier = Modifier.size(18.dp)
-                                        )
-                                    }
-                                }
-                            },
-                            actions = {
-                                // Only shown during first-launch setup (no back button yet) so the user has a
-                                // way to finish and enter the library. When opened from the library the back
-                                // button already covers this, so the redundant action is hidden.
-                                if (onBack == null) {
-                                    Box(
-                                        modifier = Modifier
-                                            .padding(end = 12.dp)
-                                            .clip(RoundedCornerShape(50))
-                                            .background(gradientBrush)
-                                            .clickable {
-                                                viewModel.saveCredentials()
-                                                viewModel.finishSetup()
-                                                onGoToLibrary()
-                                            }
-                                            .padding(horizontal = 14.dp, vertical = 8.dp),
-                                        contentAlignment = Alignment.Center
-                                    ) {
-                                        Row(verticalAlignment = Alignment.CenterVertically) {
-                                            Text(
-                                                "Library",
-                                                style = MaterialTheme.typography.labelLarge,
-                                                color = Color.White
-                                            )
-                                            Spacer(Modifier.width(4.dp))
-                                            Icon(
-                                                Icons.AutoMirrored.Filled.ArrowForward,
-                                                contentDescription = null,
-                                                tint = Color.White,
-                                                modifier = Modifier.size(14.dp)
-                                            )
-                                        }
-                                    }
-                                }
-                            },
-                            colors = TopAppBarDefaults.topAppBarColors(
-                                containerColor = MaterialTheme.colorScheme.surface
-                            )
-                        )
-                    }
-                ) { paddingValues ->
-                    Column(
-                        Modifier
-                            .fillMaxSize()
-                            .padding(paddingValues)
-                    ) {
-
-                        SettingsTabBar(
-                            tabs = visibleTabs,
-                            selected = selectedTab,
-                            onSelect = { selectedTab = it })
-
-                        // Fresh scroll state per tab so switching tabs starts at the top.
-                        key(selectedTab) {
-                            Column(
-                                modifier = Modifier
-                                    .fillMaxSize()
-                                    // Focus target for the screen: requesting this moves focus to the first
-                                    // focusable control inside, and focusGroup keeps d-pad traversal contained.
-                                    .focusRequester(contentFocus)
-                                    .focusGroup()
-                                    .verticalScroll(rememberScrollState())
-                                    .padding(horizontal = 16.dp, vertical = 12.dp),
-                                verticalArrangement = Arrangement.spacedBy(8.dp)
-                            ) {
-                                when (selectedTab) {
-                                    SettingsTab.GENERAL -> {
-                                        HomeLauncherSection(
-                                            isDefault = isDefaultHome,
-                                            onOpenSettings = ::openHomeLauncherSettings
-                                        )
-                                        Spacer(Modifier.height(4.dp))
-                                        DisplaySection(state, viewModel)
-                                        Spacer(Modifier.height(4.dp))
-                                        DualScreenSection(state, viewModel)
-                                        SystemSortSection(state, viewModel)
-                                        Spacer(Modifier.height(4.dp))
-                                        HideSystemsSection(state, viewModel)
-                                        Spacer(Modifier.height(4.dp))
-                                        BackgroundBrandingSection(
-                                            state,
-                                            viewModel,
-                                            onPickImage = {
-                                                backgroundImagePicker.launch(
-                                                    PickVisualMediaRequest(ActivityResultContracts.PickVisualMedia.ImageOnly)
-                                                )
-                                            }
-                                        )
-                                        Spacer(Modifier.height(4.dp))
-                                        FriendsToggleSection(state, viewModel)
-                                    }
-
-                                    SettingsTab.LOCKED_MODE -> LockedModeSection(
-                                        onManageAllowedGames = onManageAllowedGames,
-                                        onManageAllowedApps = onManageAllowedApps,
-                                    )
-
-                                    SettingsTab.MEDIA -> {
-                                        MediaStorageSection(
-                                            state,
-                                            onPickFolder = { mediaStoragePicker.launch(null) },
-                                            onUseDefault = viewModel::clearMediaStoragePath
-                                        )
-                                        Spacer(Modifier.height(8.dp))
-                                        // The two media sources are rarely used together, so they share one
-                                        // card with an inline segmented selector instead of stacking.
-                                        var mediaSub by rememberSaveable { mutableStateOf(0) }
-                                        SegmentedTabs(
-                                            options = listOf("ScreenScraper", "Artwork DB"),
-                                            selected = mediaSub,
-                                            onSelect = { mediaSub = it }
-                                        )
-                                        Spacer(Modifier.height(8.dp))
-                                        SettingsCard {
-                                            when (mediaSub) {
-                                                0 -> ScreenScraperBody(
-                                                    state,
-                                                    viewModel,
-                                                    onScrapeAllClick
-                                                )
-
-                                                else -> ArtworkDatabaseBody(state, viewModel)
-                                            }
-                                        }
-                                    }
-
-                                    SettingsTab.GAMES -> {
-                                        RomLibrarySection(
-                                            state, viewModel, storageVolumes,
-                                            onPickRomFolder = { folderPicker.launch(null) },
-                                            onRescanClick = onRescanClick,
-                                            onScrapeAllClick = onScrapeAllClick
-                                        )
-                                        Spacer(Modifier.height(4.dp))
-                                        AndroidGamesSection(state, viewModel)
-                                        Spacer(Modifier.height(4.dp))
-                                        SteamLibrarySection(state, viewModel)
-                                        Spacer(Modifier.height(4.dp))
-                                        EmulatorsSection(state, viewModel, onEmulatorConfigClick)
-                                    }
-
-                                    SettingsTab.RETRO_ACHIEVEMENTS -> RetroAchievementsSection(
-                                        state,
-                                        viewModel
-                                    )
-
-                                    SettingsTab.SAVE_SYNC -> SaveSyncSection()
-                                    SettingsTab.FRIENDS -> FriendsSettingsSection()
-                                }
-                                Spacer(Modifier.height(24.dp))
-                            }
-                        }
-                    }
-                }
-            }
-            if (state.showAndroidGameSelection) {
-                AndroidGameSelectionDialog(state = state, viewModel = viewModel)
-            }
-        }
-    }
-}
-
-@Composable
-private fun LockedModeSection(
-    onManageAllowedGames: () -> Unit,
-    onManageAllowedApps: () -> Unit,
-    viewModel: LockedModeSettingsViewModel = hiltViewModel(),
-) {
-    val uiState by viewModel.uiState.collectAsState()
-    val state = uiState.lockedModeState
-    val enabled = state == LockedModeState.READY || state == LockedModeState.LOCKED
-    var showLockConfirm by remember { mutableStateOf(false) }
-    var showNavigationSteps by rememberSaveable { mutableStateOf(true) }
-    var openSettingsAfterNotificationPermission by remember { mutableStateOf(false) }
-    var notificationPermissionRequested by remember { mutableStateOf(false) }
-    val pairingContext = LocalContext.current
-    val notificationPermission = rememberLauncherForActivityResult(
-        ActivityResultContracts.RequestPermission()
-    ) { granted ->
-        if (granted) viewModel.prepareEmbeddedPairingNotification()
-        if (granted && openSettingsAfterNotificationPermission) {
-            viewModel.beginEmbeddedPairingSetup()
-        }
-        openSettingsAfterNotificationPermission = false
-    }
-    val setupNeedsPairingNotification = !uiState.systemNavigationSetupProgress.paired &&
-            uiState.systemNavigationStatus in setOf(
-        SystemNavigationLockStatus.WIRELESS_DEBUGGING_REQUIRED,
-        SystemNavigationLockStatus.PAIRING_REQUIRED,
-    )
-    val showSystemNavigationSetupSteps = showNavigationSteps &&
-            shouldShowSystemNavigationSetupSteps(uiState.systemNavigationStatus)
-    LifecycleResumeEffect(
-        viewModel,
-        uiState.blockSystemNavigation,
-        setupNeedsPairingNotification,
-    ) {
-        if (uiState.blockSystemNavigation && setupNeedsPairingNotification) {
-            val needsPermission = Build.VERSION.SDK_INT >= 33 && ContextCompat.checkSelfPermission(
-                pairingContext,
-                Manifest.permission.POST_NOTIFICATIONS,
-            ) != PackageManager.PERMISSION_GRANTED
-            if (needsPermission && !notificationPermissionRequested) {
-                notificationPermissionRequested = true
-                openSettingsAfterNotificationPermission = false
-                notificationPermission.launch(Manifest.permission.POST_NOTIFICATIONS)
-            } else if (!needsPermission) {
-                viewModel.prepareEmbeddedPairingNotification()
-            }
-        }
-        onPauseOrDispose { }
-    }
-    LaunchedEffect(uiState.systemNavigationSetupProgress.developerOptionsEnabled) {
-        // as long as the developer options are not enabled, we poll
-        while (!uiState.systemNavigationSetupProgress.developerOptionsEnabled) {
-            viewModel.refreshSystemNavigationSetupProgress()
-            kotlinx.coroutines.delay(1_000)
-        }
-    }
-    LaunchedEffect(uiState.blockSystemNavigation, showSystemNavigationSetupSteps) {
-        while (uiState.blockSystemNavigation && showSystemNavigationSetupSteps) {
-            viewModel.refreshSystemNavigationSetupProgress()
-            kotlinx.coroutines.delay(1_000)
-        }
-    }
-
-    if (uiState.showSystemNavigationWarning) {
-        AlertDialog(
-            onDismissRequest = viewModel::dismissSystemNavigationWarning,
-            title = { Text("Enable system navigation blocking?") },
-            text = {
-                Text(
-                    "This lets eOr use Wireless debugging and low-level Android system " +
-                            "commands to block Home, Recents, edge navigation, and the " +
-                            "notification shade while Locked Mode is active.\n\n" +
-                            "This feature requires low-level system knowledge. Only enable it " +
-                            "if you understand what it does and know how to recover system " +
-                            "navigation if setup fails."
-                )
-            },
-            confirmButton = {
-                TextButton(onClick = viewModel::confirmBlockSystemNavigation) {
-                    Text("Enable")
-                }
-            },
-            dismissButton = {
-                TextButton(onClick = viewModel::dismissSystemNavigationWarning) {
-                    Text("Cancel")
-                }
-            },
-        )
-    }
-
-    SettingsSectionHeader("Locked mode")
-    SettingsCard {
-        Text(
-            "Locked Mode creates a simplified Home screen that shows only the games and " +
-                    "apps you choose. Enabling the feature does not lock eOr immediately—you " +
-                    "decide when to enter Locked Mode.",
-            style = MaterialTheme.typography.bodyMedium,
-        )
-        Spacer(Modifier.height(12.dp))
-        HorizontalDivider(color = MaterialTheme.colorScheme.outlineVariant)
-        Spacer(Modifier.height(12.dp))
-        Row(verticalAlignment = Alignment.CenterVertically) {
-            Icon(Icons.Default.Lock, contentDescription = null, tint = ElectricBlue)
-            Spacer(Modifier.width(12.dp))
-            Column(Modifier.weight(1f)) {
-                Text(
-                    when (state) {
-                        null -> "Loading…"
-                        LockedModeState.DISABLED -> "Disabled"
-                        LockedModeState.READY -> "Enabled"
-                        LockedModeState.LOCKED -> "Enabled and locked"
-                    },
-                    style = MaterialTheme.typography.titleMedium
-                )
-                Text(
-                    when (state) {
-                        null -> "Checking Locked Mode status…"
-                        LockedModeState.DISABLED ->
-                            "Turn on Locked Mode to restrict eOr to the games and apps you allow."
-
-                        LockedModeState.READY ->
-                            "Locked Mode is ready. Choose Lock Now here or use the lock button " +
-                                    "on Home."
-
-                        LockedModeState.LOCKED ->
-                            "Locked Mode is active. Only allowed games and apps are available."
-                    },
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            }
-            Switch(
-                checked = enabled,
-                onCheckedChange = viewModel::setEnabled,
-                enabled = state != null,
-            )
-        }
-        Spacer(Modifier.height(12.dp))
-        GradientFillButton(
-            text = "Lock Now",
-            onClick = {
-                if (uiState.hasPin) showLockConfirm = true else viewModel.lockNow()
-            },
-            modifier = Modifier.fillMaxWidth(),
-            enabled = enabled,
-        )
-        Spacer(Modifier.height(8.dp))
-        Text(
-            "Turning Locked Mode off exits it immediately. Your PIN and allowed games and " +
-                    "apps are kept for the next time you enable it.",
-            style = MaterialTheme.typography.bodySmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant,
-        )
-    }
-
-    SettingsSectionHeader("PIN settings")
-    SettingsCard {
-        Text(
-            "A PIN is optional. Set one if you want to prevent someone from leaving Locked " +
-                    "Mode without entering four digits.",
-            style = MaterialTheme.typography.bodyMedium,
-        )
-        Spacer(Modifier.height(6.dp))
-        Text(
-            when {
-                !enabled -> "Enable Locked Mode to set, change, or remove its PIN."
-                uiState.hasPin ->
-                    "PIN protection is active. Changing or removing the PIN does not " +
-                            "require the current PIN."
-
-                else ->
-                    "Without a PIN, pressing the unlock button on Home exits Locked Mode " +
-                            "immediately."
-            },
-            style = MaterialTheme.typography.bodySmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant,
-        )
-        Spacer(Modifier.height(12.dp))
-        Row(
-            horizontalArrangement = Arrangement.spacedBy(12.dp),
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            GradientOutlineButton(
-                text = if (uiState.hasPin) "Change PIN" else "Set PIN",
-                onClick = if (uiState.hasPin) viewModel::startChange else viewModel::startSetup,
-                modifier = Modifier.weight(1f),
-                enabled = enabled,
-            )
-            GradientOutlineButton(
-                text = "Remove PIN",
-                onClick = viewModel::removePin,
-                modifier = Modifier.weight(1f),
-                enabled = enabled && uiState.hasPin,
-                contentColor = MaterialTheme.colorScheme.error
-            )
-        }
-    }
-
-    // only show option to block system navigation on a developer enabled device
-    if (uiState.systemNavigationSetupProgress.developerOptionsEnabled) {
-        SettingsSectionHeader("System navigation")
-        SettingsCard {
-            Row(verticalAlignment = Alignment.CenterVertically) {
-                Icon(Icons.Default.Lock, contentDescription = null, tint = ElectricBlue)
-                Spacer(Modifier.width(12.dp))
-                Column(Modifier.weight(1f)) {
-                    Text(
-                        "Block system navigation while locked",
-                        style = MaterialTheme.typography.titleMedium
-                    )
-                    Text(
-                        "Uses Wireless debugging to block Home, Recents, edge navigation, and " +
-                                "notification shade access while Locked Mode is active.",
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    )
-                }
-                Switch(
-                    checked = uiState.blockSystemNavigation,
-                    onCheckedChange = { checked ->
-                        viewModel.setBlockSystemNavigation(checked)
-                    },
-                    enabled = enabled,
-                )
-            }
-            if (uiState.blockSystemNavigation) {
-                Spacer(Modifier.height(12.dp))
-                HorizontalDivider(color = MaterialTheme.colorScheme.outlineVariant)
-                Spacer(Modifier.height(12.dp))
-                Text(
-                    "What to expect:\n" +
-                            "• While Locked Mode is active, the notification shade, Home, Recent " +
-                            "apps, and edge navigation will be unavailable.\n" +
-                            "• Unlocking eOr restores normal system navigation.",
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                )
-                if (showSystemNavigationSetupSteps) {
-                    Spacer(Modifier.height(12.dp))
-                    val setupStepCompletion = systemNavigationSetupStepCompletion(
-                        uiState.systemNavigationSetupProgress,
-                    )
-                    val stepsText = buildAnnotatedString {
-                        append("What to do:")
-                        fun appendStep(number: Int, complete: Boolean, instruction: String) {
-                            append('\n')
-                            if (complete) append("✓")
-                            else withStyle(SpanStyle(color = Color.Transparent)) { append("✓") }
-                            append(" ($number) $instruction")
-                        }
-                        appendStep(
-                            number = 1,
-                            complete = setupStepCompletion[0],
-                            instruction = "Enable Developer options: open About phone or About " +
-                                    "device, find Build number, and tap it seven times.",
-                        )
-                        appendStep(
-                            number = 2,
-                            complete = setupStepCompletion[1],
-                            instruction = "Open Developer options and enable Wireless debugging.",
-                        )
-                        if (uiState.systemNavigationSetupProgress.paired) {
-                            appendStep(
-                                number = 3,
-                                complete = setupStepCompletion[2],
-                                instruction = "eOr is paired with Wireless debugging.",
-                            )
-                        } else {
-                            appendStep(
-                                number = 3,
-                                complete = false,
-                                instruction = "Open Wireless debugging and choose ‘Pair device " +
-                                        "with pairing code’.",
-                            )
-                            appendStep(
-                                number = 4,
-                                complete = false,
-                                instruction = "Pull down the notification shade and find the eOr " +
-                                        "notification, not leaving the pair device screen.",
-                            )
-                            appendStep(
-                                number = 5,
-                                complete = false,
-                                instruction = "Tap ‘Enter pairing code’ and enter the six-digit code.",
-                            )
-                            appendStep(
-                                number = 6,
-                                complete = false,
-                                instruction = "Press send and keep the pairing dialog open until " +
-                                        "pairing completes.",
-                            )
-                        }
-                    }
-                    Text(
-                        stepsText,
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    )
-                    Spacer(Modifier.height(8.dp))
-                    Row(
-                        modifier = Modifier.fillMaxWidth(),
-                        horizontalArrangement = Arrangement.End,
-                    ) {
-                        Text(
-                            "[hide steps]",
-                            style = MaterialTheme.typography.labelMedium,
-                            color = ElectricBlue,
-                            modifier = Modifier
-                                .dpadFocusable { showNavigationSteps = false }
-                                .padding(6.dp),
-                        )
-                    }
-                    Spacer(Modifier.height(6.dp))
-                    HorizontalDivider(color = MaterialTheme.colorScheme.outlineVariant)
-                    Spacer(Modifier.height(12.dp))
-                }
-                val statusText = when (uiState.systemNavigationStatus) {
-                    SystemNavigationLockStatus.DISABLED -> "Disabled"
-                    SystemNavigationLockStatus.UNSUPPORTED -> "Unsupported on Android 8–10"
-                    SystemNavigationLockStatus.DEVELOPER_OPTIONS_REQUIRED -> "Enable Developer options"
-                    SystemNavigationLockStatus.WIRELESS_DEBUGGING_REQUIRED ->
-                        "Enable Wireless debugging"
-
-                    SystemNavigationLockStatus.PAIRING_REQUIRED -> "Pair eOr with Wireless debugging"
-                    SystemNavigationLockStatus.DISCOVERING -> "Discovering local ADB endpoint…"
-                    SystemNavigationLockStatus.START_REQUIRED -> "Navigation blocking needs setup"
-                    SystemNavigationLockStatus.STARTING -> "Waiting for Android…"
-                    SystemNavigationLockStatus.READY -> "Ready - activates with Locked Mode"
-                    SystemNavigationLockStatus.APPLYING -> "Applying navigation blocking…"
-                    SystemNavigationLockStatus.ACTIVE -> "System navigation is blocked"
-                    SystemNavigationLockStatus.RESTORE_REQUIRED ->
-                        "Restore navigation or reboot Android"
-
-                    SystemNavigationLockStatus.ERROR -> "Could not enable navigation blocking"
-                }
-                Text(statusText, style = MaterialTheme.typography.bodyMedium)
-                val requiresDeveloperOptions = uiState.systemNavigationStatus in setOf(
-                    SystemNavigationLockStatus.DEVELOPER_OPTIONS_REQUIRED,
-                    SystemNavigationLockStatus.PAIRING_REQUIRED,
-                    SystemNavigationLockStatus.WIRELESS_DEBUGGING_REQUIRED,
-                )
-                if (requiresDeveloperOptions) {
-                    Spacer(Modifier.height(10.dp))
-                    val developerOptionsRequired =
-                        uiState.systemNavigationStatus ==
-                                SystemNavigationLockStatus.DEVELOPER_OPTIONS_REQUIRED
-                    GradientOutlineButton(
-                        text = if (developerOptionsRequired) {
-                            "Open About device"
-                        } else {
-                            "Open Developer options"
-                        },
-                        onClick = {
-                            if (developerOptionsRequired) {
-                                viewModel.openDeviceInfoSettings()
-                                return@GradientOutlineButton
-                            }
-                            val needsNotificationPermission =
-                                uiState.systemNavigationStatus ==
-                                        SystemNavigationLockStatus.PAIRING_REQUIRED &&
-                                        Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU &&
-                                        ContextCompat.checkSelfPermission(
-                                            pairingContext,
-                                            Manifest.permission.POST_NOTIFICATIONS,
-                                        ) != PackageManager.PERMISSION_GRANTED
-
-                            if (needsNotificationPermission) {
-                                notificationPermissionRequested = true
-                                openSettingsAfterNotificationPermission = true
-                                notificationPermission.launch(Manifest.permission.POST_NOTIFICATIONS)
-                            } else {
-                                viewModel.beginEmbeddedPairingSetup()
-                            }
-                        },
-                        modifier = Modifier.fillMaxWidth(),
-                    )
-                }
-            }
-        }
-    }
-
-    SettingsSectionHeader("Configuration")
-    SettingsCard {
-        Text(
-            "Build the selection that will be available on Home while Locked Mode is active.",
-            style = MaterialTheme.typography.bodyMedium,
-        )
-        Spacer(Modifier.height(6.dp))
-        Text(
-            if (enabled) "You can update the allowed selection whenever Locked Mode is enabled."
-            else "Enable Locked Mode before choosing which games and apps are allowed.",
-            style = MaterialTheme.typography.bodySmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant,
-        )
-        Spacer(Modifier.height(12.dp))
-        GradientFillButton(
-            text = "Manage allowed games",
-            onClick = onManageAllowedGames,
-            modifier = Modifier.fillMaxWidth(),
-            enabled = enabled,
-        )
-        Spacer(Modifier.height(10.dp))
-        GradientFillButton(
-            text = "Manage allowed apps",
-            onClick = onManageAllowedApps,
-            modifier = Modifier.fillMaxWidth(),
-            enabled = enabled,
-        )
-    }
-
-    if (showLockConfirm) {
-        LockedModeActivationDialog(
-            onConfirm = {
-                showLockConfirm = false
-                viewModel.lockNow()
-            },
-            onDismiss = { showLockConfirm = false },
-        )
-    }
-
-    uiState.dialogStep?.let { step ->
-        val title = when (step) {
-            LockedModeDialogStep.CREATE_PIN -> "Create PIN"
-            LockedModeDialogStep.CONFIRM_PIN -> "Confirm PIN"
-            LockedModeDialogStep.NEW_PIN -> "New PIN"
-            LockedModeDialogStep.CONFIRM_NEW_PIN -> "Confirm new PIN"
-        }
-        PinPadDialog(
-            title = title,
-            subtitle = null,
-            error = uiState.error,
-            onDismiss = viewModel::dismissDialog,
-            onPinComplete = viewModel::submitPin,
-        )
-    }
-}
-
-internal fun shouldShowSystemNavigationSetupSteps(status: SystemNavigationLockStatus): Boolean =
-    status !in setOf(
-        SystemNavigationLockStatus.READY,
-        SystemNavigationLockStatus.APPLYING,
-        SystemNavigationLockStatus.ACTIVE,
-    )
-
-internal fun systemNavigationSetupStepCompletion(
-    progress: SystemNavigationSetupProgress,
-): List<Boolean> {
-    val developerOptionsComplete = progress.developerOptionsEnabled
-    val wirelessDebuggingComplete = developerOptionsComplete && progress.wirelessDebuggingEnabled
-    val pairingComplete = wirelessDebuggingComplete && progress.paired
-    return listOf(developerOptionsComplete, wirelessDebuggingComplete, pairingComplete)
-}
-
-@Composable
-internal fun HomeLauncherSection(
-    isDefault: Boolean,
-    onOpenSettings: () -> Unit
-) {
-    SettingsSectionHeader("Home Launcher")
-    SettingsCard {
-        StatusRow(
-            icon = Icons.Default.Home,
-            text = if (isDefault) "eOr is your Home app" else "Another Home app is selected",
-            color = if (isDefault) ElectricBlue else MaterialTheme.colorScheme.onSurfaceVariant
-        )
-        Spacer(Modifier.height(12.dp))
-        if (isDefault) {
-            GradientOutlineButton(
-                text = "Manage Home app",
-                onClick = onOpenSettings,
-                modifier = Modifier.fillMaxWidth()
-            )
-        } else {
-            GradientFillButton(
-                text = "Set eOr as Home app",
-                onClick = onOpenSettings,
-                modifier = Modifier.fillMaxWidth()
-            )
-        }
-    }
-}
-
-@Composable
-private fun SettingsTabBar(
-    tabs: List<SettingsTab>,
-    selected: SettingsTab,
-    onSelect: (SettingsTab) -> Unit
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .horizontalScroll(rememberScrollState())
-            .padding(horizontal = 12.dp, vertical = 8.dp),
-        horizontalArrangement = Arrangement.spacedBy(8.dp),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        tabs.forEach { tab ->
-            val isSel = tab == selected
-            Row(
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.spacedBy(6.dp),
-                modifier = Modifier
-                    .clip(RoundedCornerShape(50))
-                    .then(
-                        if (isSel) Modifier.background(gradientBrush)
-                        else Modifier.background(MaterialTheme.colorScheme.surfaceVariant)
-                    )
-                    .dpadFocusable(shape = RoundedCornerShape(50)) { onSelect(tab) }
-                    .padding(horizontal = 14.dp, vertical = 9.dp)
-            ) {
-                Icon(
-                    tab.icon,
-                    contentDescription = null,
-                    tint = if (isSel) Color.White else MaterialTheme.colorScheme.onSurfaceVariant,
-                    modifier = Modifier.size(16.dp)
-                )
-                Text(
-                    tab.label,
-                    style = MaterialTheme.typography.labelMedium,
-                    color = if (isSel) Color.White else MaterialTheme.colorScheme.onSurfaceVariant,
-                    fontWeight = if (isSel) FontWeight.SemiBold else FontWeight.Normal
-                )
-            }
-        }
-    }
-}
-
-// ── Section: Display ──────────────────────────────────────────────────────
-
-@Composable
-private fun DisplaySection(state: SettingsUiState, viewModel: SettingsViewModel) {
-    SettingsSectionHeader("Display")
-    SettingsCard {
-        CardSwitchRow(
-            label = "Favorites tab",
-            checked = state.showFavorites,
-            onCheckedChange = viewModel::setShowFavorites
-        )
-        CardSwitchRow(
-            label = "Recently Played tab",
-            checked = state.showRecentlyPlayed,
-            onCheckedChange = viewModel::setShowRecentlyPlayed
-        )
-        CardSwitchRow(
-            label = "RetroAchievements tab",
-            checked = state.showRetroAchievements,
-            onCheckedChange = viewModel::setShowRetroAchievements
-        )
-        Spacer(Modifier.height(10.dp))
-        Text(
-            "Library Layout",
-            style = MaterialTheme.typography.bodyMedium,
-            color = MaterialTheme.colorScheme.onSurface
-        )
-        Spacer(Modifier.height(8.dp))
-        Row(
-            horizontalArrangement = Arrangement.spacedBy(8.dp),
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            BackgroundModeChip(
-                label = "Carousel",
-                selected = state.layoutMode == LayoutMode.CAROUSEL,
-                onClick = { viewModel.setLayoutMode(LayoutMode.CAROUSEL) },
-                modifier = Modifier.weight(1f)
-            )
-            BackgroundModeChip(
-                label = "Grid",
-                selected = state.layoutMode == LayoutMode.GRID,
-                onClick = { viewModel.setLayoutMode(LayoutMode.GRID) },
-                modifier = Modifier.weight(1f)
-            )
-            BackgroundModeChip(
-                label = "List",
-                selected = state.layoutMode == LayoutMode.LIST,
-                onClick = { viewModel.setLayoutMode(LayoutMode.LIST) },
-                modifier = Modifier.weight(1f)
-            )
-        }
-        // Master grid size — only meaningful (and only shown) when the grid layout is active. Sets the
-        // default tile size for every system's game grid; a system sized from its own grid overrides it.
-        if (state.layoutMode == LayoutMode.GRID) {
-            Spacer(Modifier.height(12.dp))
-            MasterGridSizeControl(
-                columns = state.masterGridColumns,
-                onSet = viewModel::setMasterGridColumns
-            )
-        }
-        Spacer(Modifier.height(12.dp))
-        Text(
-            "Appearance",
-            style = MaterialTheme.typography.bodyMedium,
-            color = MaterialTheme.colorScheme.onSurface
-        )
-        Spacer(Modifier.height(8.dp))
-        ThemePicker(selectedDark = state.darkMode, onSelect = viewModel::setDarkMode)
-
-        Spacer(Modifier.height(12.dp))
-        Text(
-            "Card colors",
-            style = MaterialTheme.typography.bodyMedium,
-            color = MaterialTheme.colorScheme.onSurface
-        )
-        Spacer(Modifier.height(8.dp))
-        CardColorSchemePicker(
-            scheme = state.cardColorScheme,
-            monoColor = state.cardMonoColor,
-            onSchemeSelected = viewModel::setCardColorScheme,
-            onMonoColorSelected = viewModel::setCardMonoColor
-        )
-
-        Spacer(Modifier.height(10.dp))
-        if (BuildConfig.LOW_POWER) {
-            // The lite build always runs reduced; there's nothing to toggle.
-            Text(
-                "Low-power build — performance optimizations are always on.",
-                style = MaterialTheme.typography.bodySmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant
-            )
-        } else {
-            CardSwitchRow(
-                label = "Performance mode",
-                checked = state.performanceMode,
-                onCheckedChange = viewModel::setPerformanceMode
-            )
-            Text(
-                "Reduces animations and delays video previews — recommended on low-power handhelds.",
-                style = MaterialTheme.typography.bodySmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant
-            )
-        }
-    }
-}
-
-/**
- * Master (default) game-grid size, shown in the Display section only while the grid layout is active.
- * The slider reads small → large (fewer columns = bigger tiles) to match the in-grid quick menu, with
- * an "Auto" stop at the far left that lets the grid fit the screen. Recents/Favorites keep their own
- * fixed sizing, and any system sized from its own grid overrides this default.
- */
-@Composable
-private fun MasterGridSizeControl(columns: Int, onSet: (Int) -> Unit) {
-    val screenWidthDp = LocalConfiguration.current.screenWidthDp
-    // Same column maths as the home grid so the steps line up with what the library actually shows.
-    val minCols = 3
-    val maxCols = maxOf(1, (screenWidthDp - 2 * 8 + 8) / (92 + 8)).coerceIn(minCols + 1, 12)
-
-    // Fold "Auto" (columns == 0) into the far-left slot; explicit sizes run smallest-tiles (maxCols)
-    // up to largest-tiles (minCols) rightward.
-    val sliderMax = (maxCols - minCols) + 1
-    val sliderValue = if (columns <= 0) 0 else (maxCols - columns.coerceIn(minCols, maxCols) + 1)
-
-    Column(Modifier.fillMaxWidth()) {
-        Row(Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.SpaceBetween) {
-            Text(
-                "Grid size",
-                style = MaterialTheme.typography.bodyMedium,
-                color = MaterialTheme.colorScheme.onSurface
-            )
-            Text(
-                if (columns <= 0) "Auto" else "$columns per row",
-                style = MaterialTheme.typography.labelMedium,
-                color = MaterialTheme.colorScheme.onSurfaceVariant
-            )
-        }
-        Slider(
-            value = sliderValue.toFloat(),
-            onValueChange = { v ->
-                val iv = v.roundToInt()
-                onSet(if (iv <= 0) 0 else maxCols - (iv - 1))
-            },
-            valueRange = 0f..sliderMax.toFloat(),
-            steps = (sliderMax - 1).coerceAtLeast(0),
-            colors = SliderDefaults.colors(
-                thumbColor = ElectricBlue,
-                activeTrackColor = ElectricBlue,
-            )
-        )
-        Row(Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.SpaceBetween) {
-            Text(
-                "Auto",
-                style = MaterialTheme.typography.labelSmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant
-            )
-            Text(
-                "Larger",
-                style = MaterialTheme.typography.labelSmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant
-            )
-        }
-        Text(
-            "Default size for every system's game grid. Resize a single system from its grid to override it.",
-            style = MaterialTheme.typography.bodySmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-    }
-}
-
-// ── Section: Dual Screen ──────────────────────────────────────────────────
-
-/**
- * Only shown on dual-screen handhelds (a second physical display is present). Lets the user turn
- * the split layout off and manually swap which panel shows the menu vs artwork if auto-detection
- * guessed wrong for an unrecognised device.
- */
-@Composable
-private fun DualScreenSection(state: SettingsUiState, viewModel: SettingsViewModel) {
-    val context = LocalContext.current
-    val hasSecondScreen = remember {
-        val dm = context.getSystemService(Context.DISPLAY_SERVICE) as? DisplayManager
-        dm?.displays?.any { it.displayId != Display.DEFAULT_DISPLAY && it.state == Display.STATE_ON } == true
-    }
-    if (!hasSecondScreen) return
-
-    SettingsSectionHeader("Dual Screen")
-    SettingsCard {
-        Text(
-            "Two screens detected. eOr shows the menu on the bottom panel and game artwork on the top.",
-            style = MaterialTheme.typography.bodySmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-        Spacer(Modifier.height(8.dp))
-        CardSwitchRow(
-            label = "Use both screens",
-            checked = state.dualScreenEnabled,
-            onCheckedChange = viewModel::setDualScreenEnabled
-        )
-        if (state.dualScreenEnabled) {
-            CardSwitchRow(
-                label = "Swap screens",
-                checked = state.dualScreenSwap,
-                onCheckedChange = viewModel::setDualScreenSwap
-            )
-            CardSwitchRow(
-                label = "Launch games on top screen",
-                checked = state.gameLaunchOnTop,
-                onCheckedChange = viewModel::setGameLaunchOnTop
-            )
-            Text(
-                "Single-screen games (PlayStation, Game Boy, etc.) open on the top panel. DS/3DS games always use both screens.",
-                style = MaterialTheme.typography.bodySmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant
-            )
-            Spacer(Modifier.height(12.dp))
-            Text(
-                "Top screen image",
-                style = MaterialTheme.typography.bodyMedium,
-                color = MaterialTheme.colorScheme.onSurface
-            )
-            Spacer(Modifier.height(8.dp))
-            Row(
-                horizontalArrangement = Arrangement.spacedBy(8.dp),
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                com.gamelaunch.frontend.ui.dualscreen.TopScreenImage.entries.forEach { opt ->
-                    BackgroundModeChip(
-                        label = opt.label,
-                        selected = state.topScreenImage == opt,
-                        onClick = { viewModel.setTopScreenImage(opt) },
-                        modifier = Modifier.weight(1f)
-                    )
-                }
-            }
-            Text(
-                "What the top panel shows while browsing a game — its logo, an in-game screenshot, or a composited mix.",
-                style = MaterialTheme.typography.bodySmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                modifier = Modifier.padding(top = 6.dp)
-            )
-        }
-    }
-    Spacer(Modifier.height(4.dp))
-}
-
-// ── Section: Sort Systems ─────────────────────────────────────────────────
-
-@Composable
-private fun SystemSortSection(state: SettingsUiState, viewModel: SettingsViewModel) {
-    SettingsSectionHeader("Sort Systems")
-    SettingsCard {
-        Text(
-            "Choose how your consoles are ordered. Pick up to two — the first is the primary sort, the second breaks ties.",
-            style = MaterialTheme.typography.bodySmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-        Spacer(Modifier.height(12.dp))
-        com.gamelaunch.frontend.domain.platform.SystemSort.entries.forEach { sort ->
-            val rank = state.systemSort.indexOf(sort)   // -1 if not selected
-            val isSel = rank >= 0
-            val disabled = !isSel && state.systemSort.size >= 2
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(vertical = 5.dp)
-                    .clip(RoundedCornerShape(10.dp))
-                    .dpadFocusable(enabled = !disabled) { viewModel.toggleSystemSort(sort) }
-                    .padding(vertical = 8.dp, horizontal = 4.dp),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Icon(
-                    if (isSel) Icons.Default.CheckCircle else Icons.Default.RadioButtonUnchecked,
-                    contentDescription = null,
-                    tint = when {
-                        isSel -> ElectricBlue
-                        disabled -> MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.4f)
-                        else -> MaterialTheme.colorScheme.onSurfaceVariant
-                    },
-                    modifier = Modifier.size(20.dp)
-                )
-                Spacer(Modifier.width(10.dp))
-                Text(
-                    sort.label,
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = if (disabled) MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.5f)
-                    else MaterialTheme.colorScheme.onSurface,
-                    fontWeight = if (isSel) FontWeight.SemiBold else FontWeight.Normal
-                )
-                Spacer(Modifier.weight(1f))
-                if (isSel) {
-                    Box(
-                        modifier = Modifier
-                            .size(22.dp)
-                            .clip(RoundedCornerShape(50))
-                            .background(gradientBrush),
-                        contentAlignment = Alignment.Center
-                    ) {
-                        Text(
-                            "${rank + 1}",
-                            style = MaterialTheme.typography.labelSmall,
-                            color = Color.White,
-                            fontWeight = FontWeight.Bold
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
-
-// ── Section: Hide Systems ─────────────────────────────────────────────────
-
-@Composable
-private fun HideSystemsSection(state: SettingsUiState, viewModel: SettingsViewModel) {
-    if (state.libraryPlatforms.isEmpty()) return
-    SettingsSectionHeader("Hide Systems")
-    SettingsCard {
-        Text(
-            "Turn a system off to hide its whole category from the home screen. Its games stay in the library and can be shown again anytime.",
-            style = MaterialTheme.typography.bodySmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-        Spacer(Modifier.height(8.dp))
-        state.libraryPlatforms.forEachIndexed { index, platformId ->
-            if (index > 0) CardDivider()
-            val displayName = com.gamelaunch.frontend.domain.platform.PlatformDefinitions
-                .byId[platformId]?.displayName ?: platformId
-            CardSwitchRow(
-                label = displayName,
-                checked = platformId !in state.hiddenPlatforms,
-                onCheckedChange = { shown -> viewModel.setPlatformHidden(platformId, !shown) }
-            )
-        }
-    }
-}
-
-// ── Section: Save Sync ─────────────────────────────────────────────────────
-
-@Composable
-private fun SaveSyncSection() {
-    val vm: SaveSyncViewModel = hiltViewModel()
-    val ui by vm.uiState.collectAsState()
-    SettingsSectionHeader("Save Sync")
-    SyncEngineCard(
-        ui,
-        onToggle = vm::setEnabled,
-        onLink = vm::linkDevice,
-        onWifiOnly = vm::setWifiOnly,
-        onChargingOnly = vm::setChargingOnly,
-        onResolveConflict = vm::resolveConflict
-    )
-    Spacer(Modifier.height(8.dp))
-    SettingsCard {
-        Text(
-            "Sync your emulator saves across devices. This shows which installed emulators eOr can " +
-                    "sync on this device — Android blocks access to some emulators' private storage.",
-            style = MaterialTheme.typography.bodySmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-        Spacer(Modifier.height(12.dp))
-        when {
-            ui.loading ->
-                LoadingStatusRow("Scanning emulators…", MaterialTheme.colorScheme.onSurfaceVariant)
-
-            ui.statuses.isEmpty() ->
-                Text(
-                    "No known emulators detected.",
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-
-            else -> ui.statuses.forEachIndexed { i, st ->
-                if (i > 0) {
-                    Spacer(Modifier.height(6.dp)); CardDivider(); Spacer(Modifier.height(6.dp))
-                }
-                SaveSyncRow(st)
-            }
-        }
-    }
-}
-
-@Composable
-private fun SyncEngineCard(
-    ui: SaveSyncViewModel.UiState,
-    onToggle: (Boolean) -> Unit,
-    onLink: (String) -> Unit,
-    onWifiOnly: (Boolean) -> Unit,
-    onChargingOnly: (Boolean) -> Unit,
-    onResolveConflict: (com.gamelaunch.frontend.data.sync.ConflictFile, Boolean) -> Unit
-) {
-    SettingsCard {
-        Row(modifier = Modifier.fillMaxWidth(), verticalAlignment = Alignment.CenterVertically) {
-            Column(Modifier.weight(1f)) {
-                Text(
-                    "Save Sync",
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = MaterialTheme.colorScheme.onSurface,
-                    fontWeight = FontWeight.SemiBold
-                )
-                Text(
-                    when {
-                        !ui.enabled -> "Off"
-                        ui.engineStarting -> "Starting…"
-                        ui.engineRunning -> "Running"
-                        else -> "Enabled"
-                    },
-                    style = MaterialTheme.typography.labelSmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            }
-            Switch(
-                checked = ui.enabled,
-                onCheckedChange = onToggle,
-                enabled = ui.engineSupported,
-                colors = SwitchDefaults.colors(
-                    checkedThumbColor = Color.White,
-                    checkedTrackColor = ElectricBlue,
-                    uncheckedThumbColor = MaterialTheme.colorScheme.outline,
-                    uncheckedTrackColor = MaterialTheme.colorScheme.surfaceVariant
-                )
-            )
-        }
-        if (!ui.engineSupported) {
-            Spacer(Modifier.height(6.dp))
-            Text(
-                "Not available for this device's processor.",
-                style = MaterialTheme.typography.labelSmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant
-            )
-        }
-        ui.engineError?.let {
-            Spacer(Modifier.height(8.dp))
-            Text(
-                it,
-                style = MaterialTheme.typography.labelSmall,
-                color = MaterialTheme.colorScheme.error
-            )
-        }
-
-        // Run conditions + backup status — shown whenever the feature is enabled.
-        if (ui.enabled) {
-            Spacer(Modifier.height(10.dp))
-            CardDivider()
-            Spacer(Modifier.height(4.dp))
-            CardSwitchRow("Only sync on Wi-Fi", ui.wifiOnly, onWifiOnly)
-            CardSwitchRow("Only sync while charging", ui.chargingOnly, onChargingOnly)
-            if (ui.backupCount >= 0) {
-                Spacer(Modifier.height(4.dp))
-                Text(
-                    if (ui.backupCount > 0)
-                        "Backed up ${ui.backupCount} save folder${if (ui.backupCount != 1) "s" else ""} before first sync."
-                    else "No existing saves needed backing up.",
-                    style = MaterialTheme.typography.labelSmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            }
-        }
-
-        // Conflicts — Syncthing keeps a copy when the same save diverges on two devices.
-        if (ui.conflicts.isNotEmpty()) {
-            Spacer(Modifier.height(12.dp))
-            CardDivider()
-            Spacer(Modifier.height(8.dp))
-            Text(
-                "Conflicts to resolve",
-                style = MaterialTheme.typography.bodyMedium,
-                color = MaterialTheme.colorScheme.onSurface,
-                fontWeight = FontWeight.SemiBold
-            )
-            ui.conflicts.forEach { conflict ->
-                Spacer(Modifier.height(8.dp))
-                Text(
-                    conflict.name,
-                    style = MaterialTheme.typography.labelSmall,
-                    color = MaterialTheme.colorScheme.onSurface
-                )
-                Text(
-                    conflict.folderLabel,
-                    style = MaterialTheme.typography.labelSmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-                Spacer(Modifier.height(4.dp))
-                Row(horizontalArrangement = Arrangement.spacedBy(8.dp)) {
-                    GradientOutlineButton(
-                        "Keep this",
-                        onClick = { onResolveConflict(conflict, true) },
-                        modifier = Modifier.weight(1f)
-                    )
-                    GradientOutlineButton(
-                        "Discard",
-                        onClick = { onResolveConflict(conflict, false) },
-                        modifier = Modifier.weight(1f)
-                    )
-                }
-            }
-        }
-
-        // Pairing — shown once the engine is up and reporting this device's ID.
-        if (ui.enabled && ui.deviceId != null) {
-            Spacer(Modifier.height(14.dp))
-            Text(
-                "Link a device",
-                style = MaterialTheme.typography.bodyMedium,
-                color = MaterialTheme.colorScheme.onSurface,
-                fontWeight = FontWeight.SemiBold
-            )
-            Text(
-                "On your other device, open Save Sync and scan this code (or paste its ID below).",
-                style = MaterialTheme.typography.labelSmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant
-            )
-            Spacer(Modifier.height(6.dp))
-            Text(
-                "Only link your own devices — a linked device can sync your save files. Never link an " +
-                    "ID someone else sent you.",
-                style = MaterialTheme.typography.labelSmall,
-                color = MaterialTheme.colorScheme.error
-            )
-            Spacer(Modifier.height(10.dp))
-            Row(modifier = Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.Center) {
-                QrCode(ui.deviceId, size = 190.dp)
-            }
-            Spacer(Modifier.height(8.dp))
-            Text(
-                ui.deviceId,
-                style = MaterialTheme.typography.labelSmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant
-            )
-            Spacer(Modifier.height(12.dp))
-            val scanLauncher = rememberLauncherForActivityResult(ScanContract()) { result ->
-                result.contents?.let(onLink)
-            }
-            GradientFillButton(
-                text = "Scan a device's QR",
-                onClick = {
-                    scanLauncher.launch(
-                        ScanOptions().apply {
-                            setOrientationLocked(false)
-                            setBeepEnabled(false)
-                            setPrompt("Point at the other eOr device's QR")
-                        }
-                    )
-                },
-                modifier = Modifier.fillMaxWidth()
-            )
-            Spacer(Modifier.height(10.dp))
-            Text(
-                "or paste its ID",
-                style = MaterialTheme.typography.labelSmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant
-            )
-            Spacer(Modifier.height(6.dp))
-            var peer by remember { mutableStateOf("") }
-            OutlinedTextField(
-                value = peer,
-                onValueChange = { peer = it },
-                label = { Text("Other device's ID") },
-                singleLine = true,
-                modifier = Modifier.fillMaxWidth(),
-                colors = OutlinedTextFieldDefaults.colors(
-                    focusedBorderColor = ElectricBlue,
-                    cursorColor = ElectricBlue
-                )
-            )
-            Spacer(Modifier.height(8.dp))
-            GradientFillButton(
-                text = "Link device",
-                onClick = { onLink(peer); peer = "" },
-                modifier = Modifier.fillMaxWidth(),
-                enabled = peer.isNotBlank()
-            )
-            ui.linkResult?.let {
-                Spacer(Modifier.height(6.dp))
-                Text(
-                    it,
-                    style = MaterialTheme.typography.labelSmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun SaveSyncRow(status: EmulatorSyncStatus) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(vertical = 4.dp),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Column(Modifier.weight(1f)) {
-            Text(
-                status.spec.displayName,
-                style = MaterialTheme.typography.bodyMedium,
-                color = MaterialTheme.colorScheme.onSurface,
-                fontWeight = FontWeight.SemiBold
-            )
-            Text(
-                status.message,
-                style = MaterialTheme.typography.labelSmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant
-            )
-        }
-        Spacer(Modifier.width(10.dp))
-        SyncStatusChip(status.readiness)
-    }
-}
-
-@Composable
-private fun SyncStatusChip(readiness: SyncReadiness) {
-    val (label, color) = when (readiness) {
-        SyncReadiness.READY -> "Ready" to Color(0xFF3FD3A6)
-        SyncReadiness.NEEDS_SETUP -> "Needs setup" to Color(0xFFFFC04D)
-        SyncReadiness.BLOCKED -> "Blocked" to MaterialTheme.colorScheme.onSurfaceVariant
-    }
-    Box(
-        modifier = Modifier
-            .clip(RoundedCornerShape(50))
-            .background(color.copy(alpha = 0.18f))
-            .padding(horizontal = 10.dp, vertical = 4.dp)
-    ) {
-        Text(
-            label,
-            style = MaterialTheme.typography.labelSmall,
-            color = color,
-            fontWeight = FontWeight.SemiBold
-        )
-    }
-}
-
-// ── Section: Background ───────────────────────────────────────────────────
-
-@Composable
-private fun BackgroundBrandingSection(
-    state: SettingsUiState,
-    viewModel: SettingsViewModel,
-    onPickImage: () -> Unit
-) {
-    SettingsSectionHeader("Background")
-    SettingsCard {
-        val hasImage = state.backgroundImagePath.isNotBlank()
-        Text(
-            "Brand your background with an image, converted to a single-colour silhouette drawn over " +
-                    "the backdrop and recoloured to match light and dark mode. With no image, the eOr " +
-                    "silhouette is used.",
-            style = MaterialTheme.typography.bodySmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-        Spacer(Modifier.height(12.dp))
-        CardSwitchRow(
-            label = "Enable custom background",
-            checked = state.backgroundImageEnabled,
-            onCheckedChange = viewModel::setBackgroundImageEnabled
-        )
-        if (state.backgroundImageEnabled) {
-            Spacer(Modifier.height(6.dp))
-            Text(
-                "Layout",
-                style = MaterialTheme.typography.bodyMedium,
-                color = MaterialTheme.colorScheme.onSurface
-            )
-            Spacer(Modifier.height(8.dp))
-            Row(
-                horizontalArrangement = Arrangement.spacedBy(8.dp),
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                BackgroundModeChip(
-                    label = "Fill",
-                    selected = state.backgroundImageMode == "FILL",
-                    onClick = { viewModel.setBackgroundImageMode("FILL") },
-                    modifier = Modifier.weight(1f)
-                )
-                BackgroundModeChip(
-                    label = "Tile",
-                    selected = state.backgroundImageMode == "TILE",
-                    onClick = { viewModel.setBackgroundImageMode("TILE") },
-                    modifier = Modifier.weight(1f)
-                )
-            }
-            Spacer(Modifier.height(10.dp))
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.SpaceBetween
-            ) {
-                Text(
-                    "Opacity",
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = MaterialTheme.colorScheme.onSurface
-                )
-                Text(
-                    "${(state.backgroundImageOpacity * 100).roundToInt()}%",
-                    style = MaterialTheme.typography.labelMedium,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            }
-            Slider(
-                value = state.backgroundImageOpacity,
-                onValueChange = viewModel::setBackgroundImageOpacity,
-                valueRange = 0.05f..1f,
-                colors = SliderDefaults.colors(
-                    thumbColor = ElectricBlue,
-                    activeTrackColor = ElectricBlue,
-                    inactiveTrackColor = MaterialTheme.colorScheme.surfaceVariant
-                )
-            )
-        }
-        Spacer(Modifier.height(12.dp))
-        GradientFillButton(
-            text = if (hasImage) "Replace image" else "Upload image",
-            onClick = onPickImage,
-            modifier = Modifier.fillMaxWidth(),
-            loading = state.convertingBackground
-        )
-        if (hasImage) {
-            Spacer(Modifier.height(8.dp))
-            GradientOutlineButton(
-                text = "Remove image (use eOr silhouette)",
-                onClick = viewModel::clearBackgroundImage,
-                modifier = Modifier.fillMaxWidth()
-            )
-        }
-    }
-}
-
-/** Pill selector used for the Fill / Tile background layout choice. */
-@Composable
-private fun BackgroundModeChip(
-    label: String,
-    selected: Boolean,
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    Box(
-        modifier = modifier
-            .height(42.dp)
-            .clip(RoundedCornerShape(21.dp))
-            .then(
-                if (selected) Modifier.background(gradientBrush)
-                else Modifier.background(MaterialTheme.colorScheme.surfaceVariant)
-            )
-            .dpadFocusable(shape = RoundedCornerShape(21.dp), onClick = onClick),
-        contentAlignment = Alignment.Center
-    ) {
-        Text(
-            label,
-            style = MaterialTheme.typography.labelLarge,
-            color = if (selected) Color.White else MaterialTheme.colorScheme.onSurfaceVariant
-        )
-    }
-}
-
-/** Two visual cards — Light / Dark — each previewing the UI, with the active one highlighted. */
-@Composable
-private fun ThemePicker(selectedDark: Boolean, onSelect: (Boolean) -> Unit) {
-    Row(horizontalArrangement = Arrangement.spacedBy(12.dp), modifier = Modifier.fillMaxWidth()) {
-        ThemeOption(
-            "Light",
-            dark = false,
-            selected = !selectedDark,
-            onClick = { onSelect(false) },
-            modifier = Modifier.weight(1f)
-        )
-        ThemeOption(
-            "Dark",
-            dark = true,
-            selected = selectedDark,
-            onClick = { onSelect(true) },
-            modifier = Modifier.weight(1f)
-        )
-    }
-}
-
-@Composable
-private fun ThemeOption(
-    label: String,
-    dark: Boolean,
-    selected: Boolean,
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    val accent = ElectricBlue
-    val bg = if (dark) Color(0xFF06091A) else Color(0xFFEDEFF4)
-    val cardColor = if (dark) Color(0xFF172044) else Color(0xFFFFFFFF)
-    val barColors = listOf(Color(0xFF7C8CFF), Color(0xFFB07BFF), Color(0xFFFF7AA8))
-    val tileShade = if (dark) 0.55f else 0f
-    Column(
-        modifier = modifier
-            .clip(RoundedCornerShape(14.dp))
-            .border(
-                width = if (selected) 2.dp else 1.dp,
-                color = if (selected) accent else MaterialTheme.colorScheme.outline,
-                shape = RoundedCornerShape(14.dp)
-            )
-            .dpadFocusable(shape = RoundedCornerShape(14.dp), onClick = onClick)
-            .padding(8.dp)
-    ) {
-        // Mini UI mock-up: background, a top accent bar and three colourful tiles.
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(74.dp)
-                .clip(RoundedCornerShape(8.dp))
-                .background(bg)
-                .padding(7.dp),
-            verticalArrangement = Arrangement.spacedBy(6.dp)
-        ) {
-            Box(
-                Modifier
-                    .fillMaxWidth(0.55f)
-                    .height(7.dp)
-                    .clip(RoundedCornerShape(50))
-                    .background(accent)
-            )
-            Row(
-                horizontalArrangement = Arrangement.spacedBy(5.dp),
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                barColors.forEach { c ->
-                    Box(
-                        Modifier
-                            .weight(1f)
-                            .height(30.dp)
-                            .clip(RoundedCornerShape(5.dp))
-                            .background(
-                                androidx.compose.ui.graphics.lerp(
-                                    c,
-                                    Color.Black,
-                                    tileShade
-                                )
-                            )
-                    )
-                }
-            }
-            Box(
-                Modifier
-                    .fillMaxWidth(0.8f)
-                    .height(6.dp)
-                    .clip(RoundedCornerShape(50))
-                    .background(cardColor)
-            )
-        }
-        Spacer(Modifier.height(8.dp))
-        Row(verticalAlignment = Alignment.CenterVertically) {
-            Icon(
-                if (selected) Icons.Default.CheckCircle else Icons.Default.RadioButtonUnchecked,
-                contentDescription = null,
-                tint = if (selected) accent else MaterialTheme.colorScheme.onSurfaceVariant,
-                modifier = Modifier.size(18.dp)
-            )
-            Spacer(Modifier.width(6.dp))
-            Text(
-                label,
-                style = MaterialTheme.typography.bodyMedium,
-                fontWeight = if (selected) FontWeight.SemiBold else FontWeight.Normal,
-                color = MaterialTheme.colorScheme.onSurface
-            )
-        }
-    }
-}
-
-/**
- * Home-card colour scheme picker: three chips (Rainbow / Black & White / Monochrome). When
- * Monochrome is active a scrollable row of colour swatches is revealed so the user can pick the
- * single hue the whole grid is tinted with. Mirrors the tile-colour logic in Glass.kt.
- */
-@Composable
-private fun CardColorSchemePicker(
-    scheme: CardColorScheme,
-    monoColor: Int,
-    onSchemeSelected: (CardColorScheme) -> Unit,
-    onMonoColorSelected: (Int) -> Unit
-) {
-    Column(Modifier.fillMaxWidth()) {
-        Row(
-            horizontalArrangement = Arrangement.spacedBy(8.dp),
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            BackgroundModeChip(
-                label = "Rainbow",
-                selected = scheme == CardColorScheme.RAINBOW,
-                onClick = { onSchemeSelected(CardColorScheme.RAINBOW) },
-                modifier = Modifier.weight(1f)
-            )
-            BackgroundModeChip(
-                label = "B & W",
-                selected = scheme == CardColorScheme.BLACK_WHITE,
-                onClick = { onSchemeSelected(CardColorScheme.BLACK_WHITE) },
-                modifier = Modifier.weight(1f)
-            )
-            BackgroundModeChip(
-                label = "Mono",
-                selected = scheme == CardColorScheme.MONOCHROME,
-                onClick = { onSchemeSelected(CardColorScheme.MONOCHROME) },
-                modifier = Modifier.weight(1f)
-            )
-        }
-        if (scheme == CardColorScheme.MONOCHROME) {
-            Spacer(Modifier.height(10.dp))
-            Text(
-                "Tile color",
-                style = MaterialTheme.typography.bodySmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant
-            )
-            Spacer(Modifier.height(6.dp))
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .horizontalScroll(rememberScrollState()),
-                horizontalArrangement = Arrangement.spacedBy(10.dp)
-            ) {
-                MonochromeSeeds.forEach { seed ->
-                    val argb = seed.toArgb()
-                    ColorSwatch(
-                        color = seed,
-                        selected = argb == monoColor,
-                        onClick = { onMonoColorSelected(argb) }
-                    )
-                }
-            }
-        }
-    }
-}
-
-/** A single circular colour swatch for the monochrome-seed picker; a ring marks the active one. */
-@Composable
-private fun ColorSwatch(color: Color, selected: Boolean, onClick: () -> Unit) {
-    Box(
-        modifier = Modifier
-            .size(38.dp)
-            .clip(CircleShape)
-            .background(color)
-            .border(
-                width = if (selected) 3.dp else 1.dp,
-                color = if (selected) MaterialTheme.colorScheme.onSurface
-                else MaterialTheme.colorScheme.outline.copy(alpha = 0.5f),
-                shape = CircleShape
-            )
-            .dpadFocusable(shape = CircleShape, onClick = onClick),
-        contentAlignment = Alignment.Center
-    ) {
-        if (selected) {
-            Icon(
-                Icons.Default.Check,
-                contentDescription = null,
-                tint = Color.White,
-                modifier = Modifier.size(20.dp)
-            )
-        }
-    }
-}
-
-// ── Section: ROM Library ──────────────────────────────────────────────────
-
-@Composable
-private fun RomLibrarySection(
-    state: SettingsUiState,
-    viewModel: SettingsViewModel,
-    storageVolumes: List<Pair<String, String>>,
-    onPickRomFolder: () -> Unit,
-    onRescanClick: () -> Unit,
-    onScrapeAllClick: () -> Unit
-) {
-    SettingsSectionHeader("ROM Library")
-    SettingsCard {
-        if (storageVolumes.size > 1) {
-            Row(
-                horizontalArrangement = Arrangement.spacedBy(8.dp),
-                modifier = Modifier.padding(bottom = 10.dp)
-            ) {
-                storageVolumes.forEach { (label, path) ->
-                    Box(
-                        modifier = Modifier
-                            .weight(1f)
-                            .clip(RoundedCornerShape(10.dp))
-                            .background(MaterialTheme.colorScheme.surfaceVariant)
-                            .dpadFocusable(shape = RoundedCornerShape(10.dp)) {
-                                viewModel.setRomRootPath(
-                                    path
-                                )
-                            }
-                            .padding(horizontal = 12.dp, vertical = 10.dp),
-                        contentAlignment = Alignment.Center
-                    ) {
-                        Text(label, style = MaterialTheme.typography.labelMedium, maxLines = 1)
-                    }
-                }
-            }
-            CardDivider()
-        }
-        Row(
-            verticalAlignment = Alignment.CenterVertically,
-            modifier = Modifier.padding(top = if (storageVolumes.size > 1) 10.dp else 0.dp)
-        ) {
-            OutlinedTextField(
-                value = state.romRootPath.ifEmpty { "Not configured" },
-                onValueChange = { viewModel.setRomRootPath(it) },
-                label = { Text("ROM Folder") },
-                modifier = Modifier.weight(1f),
-                singleLine = true,
-                colors = OutlinedTextFieldDefaults.colors(
-                    focusedBorderColor = ElectricBlue,
-                    unfocusedBorderColor = MaterialTheme.colorScheme.outline
-                )
-            )
-            Spacer(Modifier.width(8.dp))
-            IconButton(
-                onClick = onPickRomFolder,
-                modifier = Modifier
-                    .size(48.dp)
-                    .background(MaterialTheme.colorScheme.surfaceVariant, RoundedCornerShape(10.dp))
-            ) {
-                Icon(Icons.Default.FolderOpen, contentDescription = "Browse", tint = ElectricBlue)
-            }
-        }
-        Spacer(Modifier.height(4.dp))
-        Text(
-            "Tip: for SD card use /storage/XXXX-XXXX/ROMs",
-            style = MaterialTheme.typography.labelSmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-        Spacer(Modifier.height(10.dp))
-        CardDivider()
-        Spacer(Modifier.height(10.dp))
-        Row(horizontalArrangement = Arrangement.spacedBy(10.dp)) {
-            GradientOutlineButton(
-                text = "Rescan ROMs",
-                onClick = onRescanClick,
-                modifier = Modifier.weight(1f)
-            )
-            GradientFillButton(
-                text = "Scrape All",
-                onClick = onScrapeAllClick,
-                modifier = Modifier.weight(1f)
-            )
-        }
-    }
-}
-
-// ── Section: Android Games ────────────────────────────────────────────────
-
-@Composable
-private fun AndroidGamesSection(state: SettingsUiState, viewModel: SettingsViewModel) {
-    SettingsSectionHeader("Android Games")
-    SettingsCard {
-        Text(
-            "Scan installed Android games (apps tagged as games) and add them to your library.",
-            style = MaterialTheme.typography.bodySmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-        Spacer(Modifier.height(10.dp))
-        GradientFillButton(
-            text = "Scan Android Games",
-            onClick = { viewModel.scanAndroidGames() },
-            modifier = Modifier.fillMaxWidth()
-        )
-        Spacer(Modifier.height(8.dp))
-        GradientOutlineButton(
-            text = "Select Games Manually",
-            onClick = { viewModel.showAndroidGameSelection(true) },
-            modifier = Modifier.fillMaxWidth()
-        )
-        state.androidScanResult?.let { result ->
-            Spacer(Modifier.height(6.dp))
-            StatusRow(
-                icon = Icons.Default.Check,
-                text = result,
-                color = ElectricBlue
-            )
-        }
-    }
-}
-
-// ── Section: Steam / PC Games ─────────────────────────────────────────────
-
-@Composable
-private fun SteamLibrarySection(state: SettingsUiState, viewModel: SettingsViewModel) {
-    val folderPicker = rememberLauncherForActivityResult(
-        contract = ActivityResultContracts.OpenDocumentTree()
-    ) { uri ->
-        uri?.let {
-            val path = StorageUtils.resolveTreeUriToPath(it) ?: it.toString()
-            viewModel.setSteamLibraryPath(path)
-        }
-    }
-
-    SettingsSectionHeader("Steam / PC Games")
-    SettingsCard {
-        Text(
-            "Add PC games run through GameNative, GameHub or Winlator. In GameNative, turn on " +
-                    "\"Export for ES-DE\" and point it at a folder on shared storage, then choose that same " +
-                    "folder here — eOr reads the exported games and launches them back through GameNative. " +
-                    "A raw Steam \"steamapps\" folder works too.",
-            style = MaterialTheme.typography.bodySmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-        Spacer(Modifier.height(10.dp))
-        Row(verticalAlignment = Alignment.CenterVertically) {
-            Icon(
-                Icons.Default.FolderOpen,
-                contentDescription = null,
-                tint = ElectricBlue,
-                modifier = Modifier.size(20.dp)
-            )
-            Spacer(Modifier.width(8.dp))
-            Text(
-                text = state.steamLibraryPath.ifBlank { "No folder selected" },
-                style = MaterialTheme.typography.bodyMedium,
-                color = MaterialTheme.colorScheme.onSurface,
-                maxLines = 2,
-                modifier = Modifier.weight(1f)
-            )
-        }
-        Spacer(Modifier.height(10.dp))
-        Row(horizontalArrangement = Arrangement.spacedBy(10.dp)) {
-            GradientFillButton(
-                text = if (state.steamLibraryPath.isBlank()) "Choose Folder" else "Change Folder",
-                onClick = { folderPicker.launch(null) },
-                modifier = Modifier.weight(1f)
-            )
-            if (state.steamLibraryPath.isNotBlank()) {
-                GradientOutlineButton(
-                    text = "Clear",
-                    onClick = { viewModel.clearSteamLibraryPath() },
-                    modifier = Modifier.weight(1f)
-                )
-            }
-        }
-        if (state.steamLibraryPath.isNotBlank()) {
-            Spacer(Modifier.height(8.dp))
-            GradientFillButton(
-                text = "Scan Steam Library",
-                onClick = { viewModel.scanSteamLibrary() },
-                modifier = Modifier.fillMaxWidth()
-            )
-        }
-        when {
-            state.steamScanning -> {
-                Spacer(Modifier.height(8.dp))
-                LoadingStatusRow(
-                    "Scanning Steam library…",
-                    MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            }
-
-            state.steamScanResult != null -> {
-                Spacer(Modifier.height(6.dp))
-                StatusRow(Icons.Default.Check, state.steamScanResult, ElectricBlue)
-            }
-        }
-    }
-}
-
-@Composable
-private fun AndroidGameSelectionDialog(
-    state: SettingsUiState,
-    viewModel: SettingsViewModel
-) {
-    Dialog(
-        onDismissRequest = { viewModel.showAndroidGameSelection(false) },
-        properties = DialogProperties(usePlatformDefaultWidth = false)
-    ) {
-        Surface(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(16.dp),
-            shape = RoundedCornerShape(24.dp),
-            color = MaterialTheme.colorScheme.surface,
-            tonalElevation = 6.dp
-        ) {
-            Column(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .padding(20.dp)
-            ) {
-                // Header
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.SpaceBetween,
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Column {
-                        Text(
-                            text = "Select Android Games",
-                            style = MaterialTheme.typography.titleLarge,
-                            fontWeight = FontWeight.Bold,
-                            color = MaterialTheme.colorScheme.onSurface
-                        )
-                        Spacer(Modifier.height(4.dp))
-                        Text(
-                            text = "Check the apps you want to see in your Games section",
-                            style = MaterialTheme.typography.bodySmall,
-                            color = MaterialTheme.colorScheme.onSurfaceVariant
-                        )
-                    }
-                    IconButton(
-                        onClick = { viewModel.showAndroidGameSelection(false) },
-                        modifier = Modifier
-                            .size(40.dp)
-                            .background(MaterialTheme.colorScheme.surfaceVariant, CircleShape)
-                    ) {
-                        Icon(
-                            imageVector = Icons.Default.Close,
-                            contentDescription = "Close",
-                            tint = MaterialTheme.colorScheme.onSurfaceVariant
-                        )
-                    }
-                }
-
-                Spacer(Modifier.height(16.dp))
-
-                // Scrollable list of apps
-                Box(modifier = Modifier.weight(1f)) {
-                    if (state.installedApps.isEmpty()) {
-                        Box(Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
-                            CircularProgressIndicator(color = ElectricBlue)
-                        }
-                    } else {
-                        LazyColumn(
-                            verticalArrangement = Arrangement.spacedBy(8.dp),
-                            modifier = Modifier.fillMaxSize()
-                        ) {
-                            items(state.installedApps, key = { it.packageName }) { app ->
-                                val isChecked = state.checkedPackages.contains(app.packageName)
-                                var isFocused by remember { mutableStateOf(false) }
-
-                                Row(
-                                    modifier = Modifier
-                                        .fillMaxWidth()
-                                        .clip(RoundedCornerShape(12.dp))
-                                        .background(
-                                            if (isFocused) MaterialTheme.colorScheme.surfaceVariant
-                                            else Color.Transparent
-                                        )
-                                        .border(
-                                            width = if (isFocused) 2.dp else 1.dp,
-                                            color = if (isFocused) ElectricBlue else MaterialTheme.colorScheme.outlineVariant,
-                                            shape = RoundedCornerShape(12.dp)
-                                        )
-                                        .onFocusChanged { isFocused = it.isFocused }
-                                        .focusable()
-                                        .clickable {
-                                            viewModel.toggleAndroidGameSelection(app, !isChecked)
-                                        }
-                                        .padding(horizontal = 12.dp, vertical = 10.dp),
-                                    verticalAlignment = Alignment.CenterVertically,
-                                    horizontalArrangement = Arrangement.SpaceBetween
-                                ) {
-                                    Row(
-                                        verticalAlignment = Alignment.CenterVertically,
-                                        modifier = Modifier.weight(1f)
-                                    ) {
-                                        AppIcon(
-                                            packageName = app.packageName,
-                                            packageManagerHelper = viewModel.packageManagerHelper,
-                                            modifier = Modifier.size(36.dp)
-                                        )
-                                        Spacer(Modifier.width(12.dp))
-                                        Column {
-                                            Text(
-                                                text = app.label,
-                                                style = MaterialTheme.typography.bodyMedium,
-                                                fontWeight = FontWeight.SemiBold,
-                                                color = MaterialTheme.colorScheme.onSurface,
-                                                maxLines = 1,
-                                                overflow = TextOverflow.Ellipsis
-                                            )
-                                            Text(
-                                                text = app.packageName,
-                                                style = MaterialTheme.typography.bodySmall,
-                                                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                                                maxLines = 1,
-                                                overflow = TextOverflow.Ellipsis
-                                            )
-                                        }
-                                    }
-                                    Checkbox(
-                                        checked = isChecked,
-                                        onCheckedChange = null, // Handled by row clickable
-                                        colors = CheckboxDefaults.colors(
-                                            checkedColor = ElectricBlue,
-                                            checkmarkColor = Color.White
-                                        )
-                                    )
-                                }
-                            }
-                        }
-                    }
-                }
-
-                Spacer(Modifier.height(16.dp))
-
-                // Done Button
-                GradientFillButton(
-                    text = "Done",
-                    onClick = { viewModel.showAndroidGameSelection(false) },
-                    modifier = Modifier.fillMaxWidth()
-                )
-            }
-        }
-    }
-}
-
-// ── Section: Emulators ────────────────────────────────────────────────────
-
-@Composable
-private fun EmulatorsSection(
-    state: SettingsUiState,
-    viewModel: SettingsViewModel,
-    onEmulatorConfigClick: () -> Unit
-) {
-    SettingsSectionHeader("Emulators")
-    SettingsCard {
-        Text(
-            "Auto-detect maps installed emulators to your platforms, or configure each one manually.",
-            style = MaterialTheme.typography.bodySmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-        Spacer(Modifier.height(10.dp))
-        GradientFillButton(
-            text = "Auto-detect Emulators",
-            onClick = { viewModel.autoDetectEmulators() },
-            enabled = !state.emulatorDetecting,
-            loading = state.emulatorDetecting,
-            modifier = Modifier.fillMaxWidth()
-        )
-        Spacer(Modifier.height(8.dp))
-        GradientOutlineButton(
-            text = "Configure Emulators Manually",
-            onClick = onEmulatorConfigClick,
-            modifier = Modifier.fillMaxWidth()
-        )
-    }
-    Spacer(Modifier.height(10.dp))
-    EmulatorUpdatesCard(
-        obtainiumInstalled = state.obtainiumInstalled,
-        isChecking = state.isCheckingUpdates,
-        updates = state.emulatorUpdates,
-        notificationsEnabled = state.emulatorUpdateNotifications,
-        onTrack = { viewModel.trackWithObtainium() },
-        onCheck = { viewModel.checkForEmulatorUpdates() },
-        onUpdate = { viewModel.updateWithObtainium(it) },
-        onNotificationsChange = { viewModel.setEmulatorUpdateNotifications(it) }
-    )
-}
-
-@Composable
-private fun EmulatorUpdatesCard(
-    obtainiumInstalled: Boolean,
-    isChecking: Boolean,
-    updates: List<EmulatorUpdate>,
-    notificationsEnabled: Boolean,
-    onTrack: () -> Unit,
-    onCheck: () -> Unit,
-    onUpdate: (EmulatorUpdate) -> Unit,
-    onNotificationsChange: (Boolean) -> Unit
-) {
-    SettingsCard {
-        Text("Emulator updates", style = MaterialTheme.typography.titleSmall)
-        Spacer(Modifier.height(4.dp))
-        Text(
-            if (obtainiumInstalled) {
-                "Obtainium keeps your emulators up to date in the background."
-            } else {
-                "Track and install emulator updates with Obtainium — a free, open-source app " +
-                    "updater. Tap below to install it, then come back."
-            },
-            style = MaterialTheme.typography.bodySmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-        Spacer(Modifier.height(10.dp))
-        GradientFillButton(
-            text = if (obtainiumInstalled) "Track updates" else "Set up Obtainium",
-            onClick = onTrack,
-            modifier = Modifier.fillMaxWidth()
-        )
-        Spacer(Modifier.height(8.dp))
-        GradientOutlineButton(
-            text = if (isChecking) "Checking…" else "Check for updates",
-            onClick = onCheck,
-            enabled = !isChecking,
-            modifier = Modifier.fillMaxWidth()
-        )
-        Spacer(Modifier.height(4.dp))
-        CardSwitchRow(
-            label = "Update notifications",
-            checked = notificationsEnabled,
-            onCheckedChange = onNotificationsChange
-        )
-
-        if (updates.isNotEmpty()) {
-            Spacer(Modifier.height(12.dp))
-            Text(
-                "${updates.size} update${if (updates.size != 1) "s" else ""} available",
-                style = MaterialTheme.typography.labelLarge,
-                color = ElectricBlue
-            )
-            updates.forEach { update ->
-                Spacer(Modifier.height(8.dp))
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Column(Modifier.weight(1f)) {
-                        Text(update.displayName, style = MaterialTheme.typography.bodyMedium)
-                        Text(
-                            "${update.installedVersion ?: "?"} → ${update.latestVersion}",
-                            style = MaterialTheme.typography.bodySmall,
-                            color = MaterialTheme.colorScheme.onSurfaceVariant
-                        )
-                    }
-                    Spacer(Modifier.width(12.dp))
-                    GradientFillButton(
-                        text = "Update",
-                        onClick = { onUpdate(update) },
-                        modifier = Modifier.width(104.dp)
-                    )
-                }
-            }
-        }
-    }
-}
-
-// ── Section: ScreenScraper ────────────────────────────────────────────────
-
-@Composable
-private fun ScreenScraperBody(
-    state: SettingsUiState,
-    viewModel: SettingsViewModel,
-    onScrapeAllClick: () -> Unit
-) {
-    Text(
-        "ScreenScraper is the default scraper — it provides the best box art, screenshots, marquees (wheel logos), and video previews. " +
-                "Create a free account at screenscraper.fr, then enter your username and password below. " +
-                "Without credentials the app falls back to libretro thumbnails and LaunchBox.",
-        style = MaterialTheme.typography.bodySmall,
-        color = MaterialTheme.colorScheme.onSurfaceVariant
-    )
-    Spacer(Modifier.height(12.dp))
-    OutlinedTextField(
-        value = state.ssId,
-        onValueChange = viewModel::updateSsId,
-        label = { Text("Username (ssid)") },
-        modifier = Modifier.fillMaxWidth(),
-        singleLine = true,
-        colors = OutlinedTextFieldDefaults.colors(
-            focusedBorderColor = ElectricBlue,
-            unfocusedBorderColor = MaterialTheme.colorScheme.outline
-        )
-    )
-    Spacer(Modifier.height(8.dp))
-    OutlinedTextField(
-        value = state.ssPassword,
-        onValueChange = viewModel::updateSsPassword,
-        label = { Text("Password") },
-        visualTransformation = PasswordVisualTransformation(),
-        modifier = Modifier.fillMaxWidth(),
-        singleLine = true,
-        colors = OutlinedTextFieldDefaults.colors(
-            focusedBorderColor = ElectricBlue,
-            unfocusedBorderColor = MaterialTheme.colorScheme.outline
-        )
-    )
-    Spacer(Modifier.height(10.dp))
-    Row(horizontalArrangement = Arrangement.spacedBy(10.dp)) {
-        GradientOutlineButton(
-            text = "Save",
-            onClick = { viewModel.saveCredentials() },
-            modifier = Modifier.weight(1f)
-        )
-        GradientFillButton(
-            text = "Validate",
-            onClick = { viewModel.validateCredentials() },
-            enabled = !state.credentialValidating,
-            loading = state.credentialValidating,
-            modifier = Modifier.weight(1f)
-        )
-    }
-    state.credentialValid?.let { valid ->
-        Spacer(Modifier.height(8.dp))
-        StatusRow(
-            icon = if (valid) Icons.Default.Check else Icons.Default.Close,
-            text = if (valid) "Credentials valid" else "Invalid credentials",
-            color = if (valid) ElectricBlue else MaterialTheme.colorScheme.error
-        )
-    }
-
-    Spacer(Modifier.height(12.dp))
-    CardDivider()
-    Spacer(Modifier.height(12.dp))
-
-    Text(
-        "Scrape options",
-        style = MaterialTheme.typography.labelLarge,
-        color = MaterialTheme.colorScheme.primary,
-        fontWeight = FontWeight.SemiBold
-    )
-    Spacer(Modifier.height(4.dp))
-    CardSwitchRow("Metadata", state.scrapeMetadata, viewModel::setScrapeMetadata)
-    CardSwitchRow("Box Art", state.scrapeBoxArt, viewModel::setScrapeBoxArt)
-    CardSwitchRow("Screenshots", state.scrapeScreenshots, viewModel::setScrapeScreenshots)
-    CardSwitchRow("Marquees (Wheel Logos)", state.scrapeWheelLogos, viewModel::setScrapeWheelLogos)
-    CardSwitchRow("Video Previews", state.scrapeVideos, viewModel::setScrapeVideos)
-
-    Spacer(Modifier.height(12.dp))
-    CardDivider()
-    Spacer(Modifier.height(12.dp))
-
-    Text(
-        "Scrapes every game that hasn't been scraped yet, using ScreenScraper first and falling back to free sources.",
-        style = MaterialTheme.typography.bodySmall,
-        color = MaterialTheme.colorScheme.onSurfaceVariant
-    )
-    Spacer(Modifier.height(10.dp))
-    GradientFillButton(
-        text = "Scrape Now",
-        onClick = onScrapeAllClick,
-        modifier = Modifier.fillMaxWidth()
-    )
-}
-
-// ── Section: Artwork Database ─────────────────────────────────────────────
-
-@Composable
-private fun ArtworkDatabaseBody(state: SettingsUiState, viewModel: SettingsViewModel) {
-    Text(
-        "LaunchBox DB — box art & screenshots. ~190 MB, one-time download. No account needed.",
-        style = MaterialTheme.typography.bodySmall,
-        color = MaterialTheme.colorScheme.onSurfaceVariant
-    )
-    Spacer(Modifier.height(10.dp))
-
-    val lbSyncing = state.lbSyncStatus is LbSyncStatus.Downloading ||
-            state.lbSyncStatus is LbSyncStatus.Parsing
-
-    GradientFillButton(
-        text = if (lbSyncing) "Syncing…" else "Sync Artwork DB",
-        onClick = { viewModel.syncLaunchBox() },
-        enabled = !lbSyncing,
-        modifier = Modifier.fillMaxWidth(),
-        loading = lbSyncing
-    )
-
-    when (val status = state.lbSyncStatus) {
-        is LbSyncStatus.Downloading -> {
-            Spacer(Modifier.height(8.dp))
-            LinearProgressIndicator(
-                modifier = Modifier.fillMaxWidth(),
-                color = ElectricBlue
-            )
-            Spacer(Modifier.height(4.dp))
-            Text(
-                "Downloading database…",
-                style = MaterialTheme.typography.labelSmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant
-            )
-        }
-
-        is LbSyncStatus.Parsing -> {
-            Spacer(Modifier.height(8.dp))
-            LinearProgressIndicator(
-                modifier = Modifier.fillMaxWidth(),
-                color = NeonPurple
-            )
-            Spacer(Modifier.height(4.dp))
-            Text(
-                "Parsing… ${"%,d".format(status.gamesIndexed)} games indexed",
-                style = MaterialTheme.typography.labelSmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant
-            )
-        }
-
-        is LbSyncStatus.Complete -> {
-            Spacer(Modifier.height(6.dp))
-            StatusRow(
-                icon = Icons.Default.Check,
-                text = "Sync complete — ${"%,d".format(status.totalGames)} games",
-                color = ElectricBlue
-            )
-        }
-
-        is LbSyncStatus.Error -> {
-            Spacer(Modifier.height(6.dp))
-            StatusRow(
-                icon = Icons.Default.Close,
-                text = status.message,
-                color = MaterialTheme.colorScheme.error
-            )
-        }
-
-        null -> {
-            if (state.lbGameCount > 0) {
-                Spacer(Modifier.height(6.dp))
-                Text(
-                    "${"%,d".format(state.lbGameCount)} games in local database",
-                    style = MaterialTheme.typography.labelSmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            }
-        }
-    }
-}
-
-// ── Section: RetroAchievements ────────────────────────────────────────────
-
-@Composable
-private fun RetroAchievementsSection(state: SettingsUiState, viewModel: SettingsViewModel) {
-    SettingsSectionHeader("RetroAchievements")
-    SettingsCard {
-        Text(
-            "Sign in with your RetroAchievements username and password to see your points and profile.",
-            style = MaterialTheme.typography.bodySmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-        Spacer(Modifier.height(12.dp))
-        OutlinedTextField(
-            value = state.raUsername,
-            onValueChange = viewModel::updateRaUsername,
-            label = { Text("Username") },
-            modifier = Modifier.fillMaxWidth(),
-            singleLine = true,
-            colors = OutlinedTextFieldDefaults.colors(
-                focusedBorderColor = ElectricBlue,
-                unfocusedBorderColor = MaterialTheme.colorScheme.outline
-            )
-        )
-        Spacer(Modifier.height(8.dp))
-        OutlinedTextField(
-            value = state.raPassword,
-            onValueChange = viewModel::updateRaPassword,
-            label = { Text("Password") },
-            visualTransformation = PasswordVisualTransformation(),
-            modifier = Modifier.fillMaxWidth(),
-            singleLine = true,
-            colors = OutlinedTextFieldDefaults.colors(
-                focusedBorderColor = ElectricBlue,
-                unfocusedBorderColor = MaterialTheme.colorScheme.outline
-            )
-        )
-        Spacer(Modifier.height(10.dp))
-        GradientFillButton(
-            text = if (state.raLoggedIn) "Update Sign-In" else "Sign In",
-            onClick = { viewModel.saveRaCredentials() },
-            enabled = !state.raLoggingIn,
-            loading = state.raLoggingIn,
-            modifier = Modifier.fillMaxWidth()
-        )
-        state.raLoginResult?.let { msg ->
-            val ok = msg.startsWith("Signed in")
-            Spacer(Modifier.height(8.dp))
-            StatusRow(
-                icon = if (ok) Icons.Default.Check else Icons.Default.Close,
-                text = msg,
-                color = if (ok) ElectricBlue else MaterialTheme.colorScheme.error
-            )
-        }
-        if (state.raLoggedIn) {
-            Spacer(Modifier.height(6.dp))
-            GradientOutlineButton(
-                text = "Sign Out",
-                onClick = { viewModel.signOutRa() },
-                modifier = Modifier.fillMaxWidth()
-            )
-        }
-
-        Spacer(Modifier.height(12.dp))
-        CardDivider()
-        Spacer(Modifier.height(12.dp))
-
-        Text(
-            "Optional: add a Web API Key (retroachievements.org → Settings → Keys) to also see your rank and recently-played games with completion progress.",
-            style = MaterialTheme.typography.bodySmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-        Spacer(Modifier.height(8.dp))
-        OutlinedTextField(
-            value = state.raApiKey,
-            onValueChange = viewModel::updateRaApiKey,
-            label = { Text("Web API Key (optional)") },
-            visualTransformation = PasswordVisualTransformation(),
-            modifier = Modifier.fillMaxWidth(),
-            singleLine = true,
-            colors = OutlinedTextFieldDefaults.colors(
-                focusedBorderColor = ElectricBlue,
-                unfocusedBorderColor = MaterialTheme.colorScheme.outline
-            )
-        )
-    }
-}
-
-// ── Reusable design system components ─────────────────────────────────────
-
-/** Inline segmented control — connected pill segments, the selected one filled. */
-@Composable
-private fun SegmentedTabs(
-    options: List<String>,
-    selected: Int,
-    onSelect: (Int) -> Unit
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clip(RoundedCornerShape(12.dp))
-            .background(MaterialTheme.colorScheme.surfaceVariant)
-            .padding(4.dp),
-        horizontalArrangement = Arrangement.spacedBy(4.dp)
-    ) {
-        options.forEachIndexed { i, label ->
-            val isSel = i == selected
-            Box(
-                modifier = Modifier
-                    .weight(1f)
-                    .clip(RoundedCornerShape(9.dp))
-                    .then(if (isSel) Modifier.background(gradientBrush) else Modifier)
-                    .dpadFocusable(shape = RoundedCornerShape(9.dp)) { onSelect(i) }
-                    .padding(vertical = 9.dp),
-                contentAlignment = Alignment.Center
-            ) {
-                Text(
-                    label,
-                    style = MaterialTheme.typography.labelMedium,
-                    color = if (isSel) Color.White else MaterialTheme.colorScheme.onSurfaceVariant,
-                    fontWeight = if (isSel) FontWeight.SemiBold else FontWeight.Normal,
-                    maxLines = 1
-                )
-            }
-        }
-    }
-}
-
-// ── Section: Media Storage ────────────────────────────────────────────────
-
-@Composable
-private fun MediaStorageSection(
-    state: SettingsUiState,
-    onPickFolder: () -> Unit,
-    onUseDefault: () -> Unit
-) {
-    SettingsSectionHeader("Media Storage")
-    SettingsCard {
-        Text(
-            "Choose where scraped box art, screenshots and videos are saved — for example your SD card. " +
-                    "Optional: if you don't pick a folder, media is kept in the app's internal storage. " +
-                    "If the folder already contains media (e.g. an ES-DE library), it's imported automatically.",
-            style = MaterialTheme.typography.bodySmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-        Spacer(Modifier.height(10.dp))
-        Row(verticalAlignment = Alignment.CenterVertically) {
-            Icon(
-                Icons.Default.FolderOpen,
-                contentDescription = null,
-                tint = ElectricBlue,
-                modifier = Modifier.size(20.dp)
-            )
-            Spacer(Modifier.width(8.dp))
-            Text(
-                text = state.mediaStoragePath.ifBlank { "Default — internal storage" },
-                style = MaterialTheme.typography.bodyMedium,
-                color = MaterialTheme.colorScheme.onSurface,
-                maxLines = 2,
-                modifier = Modifier.weight(1f)
-            )
-        }
-        Spacer(Modifier.height(10.dp))
-        Row(horizontalArrangement = Arrangement.spacedBy(10.dp)) {
-            GradientFillButton(
-                text = "Choose Folder",
-                onClick = onPickFolder,
-                modifier = Modifier.weight(1f)
-            )
-            if (state.mediaStoragePath.isNotBlank()) {
-                GradientOutlineButton(
-                    text = "Use Default",
-                    onClick = onUseDefault,
-                    modifier = Modifier.weight(1f)
-                )
-            }
-        }
-
-        // Auto-import feedback for the chosen folder.
-        when (val s = state.esdeImportStatus) {
-            is EsdeImportStatus.Scanning -> {
-                Spacer(Modifier.height(8.dp))
-                LoadingStatusRow(
-                    "Checking folder for existing media…",
-                    MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            }
-
-            is EsdeImportStatus.Complete -> {
-                Spacer(Modifier.height(8.dp))
-                if (s.matched > 0) {
-                    StatusRow(
-                        Icons.Default.Check,
-                        "Imported media for ${s.matched} game${if (s.matched == 1) "" else "s"}",
-                        ElectricBlue
-                    )
-                } else {
-                    StatusRow(
-                        Icons.Default.Check,
-                        "No existing media found — new scrapes will be saved here",
-                        MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-                }
-            }
-
-            is EsdeImportStatus.Error -> {
-                Spacer(Modifier.height(8.dp))
-                StatusRow(Icons.Default.Close, s.message, MaterialTheme.colorScheme.error)
-            }
-
-            else -> {}
-        }
-    }
-}
-
-@Composable
-private fun SettingsSectionHeader(title: String) {
-    Text(
-        text = title,
-        style = MaterialTheme.typography.labelLarge,
-        color = ElectricBlue,
-        modifier = Modifier.padding(start = 4.dp, bottom = 4.dp)
-    )
-}
-
-@Composable
-private fun SettingsCard(content: @Composable () -> Unit) {
-    Card(
-        modifier = Modifier.fillMaxWidth(),
-        shape = RoundedCornerShape(16.dp),
-        colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.surface),
-        elevation = CardDefaults.cardElevation(0.dp)
-    ) {
-        Column(modifier = Modifier.padding(16.dp)) {
-            content()
-        }
-    }
-}
-
-@Composable
-private fun CardDivider() {
-    HorizontalDivider(
-        color = MaterialTheme.colorScheme.outline.copy(alpha = 0.6f),
-        thickness = 0.5.dp
-    )
-}
-
-@Composable
-private fun CardSwitchRow(
-    label: String,
-    checked: Boolean,
-    onCheckedChange: (Boolean) -> Unit
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clip(RoundedCornerShape(8.dp))
-            .dpadFocusable(shape = RoundedCornerShape(8.dp)) { onCheckedChange(!checked) }
-            .padding(horizontal = 4.dp, vertical = 6.dp),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.SpaceBetween
-    ) {
-        Text(label, style = MaterialTheme.typography.bodyMedium)
-        Switch(
-            checked = checked,
-            onCheckedChange = onCheckedChange,
-            colors = SwitchDefaults.colors(
-                checkedThumbColor = Color.White,
-                checkedTrackColor = ElectricBlue,
-                uncheckedThumbColor = MaterialTheme.colorScheme.outline,
-                uncheckedTrackColor = MaterialTheme.colorScheme.surfaceVariant
-            )
-        )
-    }
-}
-
-@Composable
-private fun GradientFillButton(
-    text: String,
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier,
-    enabled: Boolean = true,
-    loading: Boolean = false
-) {
-    val alpha = if (enabled) 1f else 0.5f
-    Box(
-        modifier = modifier
-            .height(46.dp)
-            .clip(RoundedCornerShape(23.dp))
-            .background(
-                Brush.horizontalGradient(
-                    listOf(
-                        ElectricBlue.copy(alpha = alpha),
-                        NeonPurple.copy(alpha = alpha)
-                    )
-                )
-            )
-            .then(
-                if (enabled) Modifier.dpadFocusable(
-                    shape = RoundedCornerShape(23.dp),
-                    onClick = onClick
-                ) else Modifier
-            ),
-        contentAlignment = Alignment.Center
-    ) {
-        if (loading) {
-            CircularProgressIndicator(
-                modifier = Modifier.size(20.dp),
-                color = Color.White,
-                strokeWidth = 2.dp
-            )
-        } else {
-            Text(text, style = MaterialTheme.typography.labelLarge, color = Color.White)
-        }
-    }
-}
-
-@Composable
-private fun GradientOutlineButton(
-    text: String,
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier,
-    enabled: Boolean = true,
-    contentColor: Color = MaterialTheme.colorScheme.primary
-) {
-    Box(
-        modifier = modifier
-            .height(46.dp)
-            .clip(RoundedCornerShape(23.dp))
-            .background(MaterialTheme.colorScheme.surfaceVariant)
-            .then(
-                if (enabled) Modifier.dpadFocusable(
-                    shape = RoundedCornerShape(23.dp),
-                    onClick = onClick
-                ) else Modifier
-            ),
-        contentAlignment = Alignment.Center
-    ) {
-        Text(
-            text = text,
-            style = MaterialTheme.typography.labelLarge,
-            color = if (enabled) contentColor
-            else MaterialTheme.colorScheme.onSurfaceVariant
-        )
-    }
-}
-
-@Composable
-private fun StatusRow(
-    icon: ImageVector,
-    text: String,
-    color: Color
-) {
-    Row(verticalAlignment = Alignment.CenterVertically) {
-        Icon(icon, contentDescription = null, tint = color, modifier = Modifier.size(16.dp))
-        Spacer(Modifier.width(6.dp))
-        Text(text, style = MaterialTheme.typography.labelSmall, color = color)
-    }
-}
-
-@Composable
-private fun LoadingStatusRow(text: String, color: Color) {
-    Row(verticalAlignment = Alignment.CenterVertically) {
-        CircularProgressIndicator(
-            color = ElectricBlue,
-            strokeWidth = 2.dp,
-            modifier = Modifier.size(16.dp)
-        )
-        Spacer(Modifier.width(6.dp))
-        Text(text, style = MaterialTheme.typography.labelSmall, color = color)
-    }
-}
-
-// ── Section: Friends ──────────────────────────────────────────────────────
-
-/** The master opt-in/out toggle, shown in the General tab so it's always reachable. */
-@Composable
-private fun FriendsToggleSection(state: SettingsUiState, viewModel: SettingsViewModel) {
-    SettingsSectionHeader("Friends")
-    SettingsCard {
-        Text(
-            "See a friend's last-played game and RetroAchievements score. Peer-to-peer — no account, " +
-                    "nothing stored online. Turning this off stops all sharing and hides the Friends tab on the home screen.",
-            style = MaterialTheme.typography.labelSmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-        Spacer(Modifier.height(8.dp))
-        CardSwitchRow("Enable Friends", state.friendsEnabled, viewModel::setFriendsEnabled)
-    }
-}
-
-/** Pairing + friends-management, shown as its own Settings tab (only while Friends is enabled). */
-@Composable
-private fun FriendsSettingsSection() {
-    val vm: FriendsViewModel = hiltViewModel()
-    val ui by vm.uiState.collectAsState()
-    val context = LocalContext.current
-    var codeInput by rememberSaveable { mutableStateOf("") }
-
-    if (!ui.engineSupported) {
-        SettingsSectionHeader("Friends")
-        SettingsCard { Text("Friends needs the sync engine, which isn't available on this device.") }
-        return
-    }
-
-    // Master switch — always shown here so this tab stays reachable. Turning it off stops all
-    // sharing and hides the Home Friends tab, but keeps this tab (and your friends list).
-    SettingsSectionHeader("Friends")
-    SettingsCard {
-        Text(
-            "Turning this off stops all sharing and hides the Friends tab on the home screen. Your friends list is kept for when you turn it back on.",
-            style = MaterialTheme.typography.labelSmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-        Spacer(Modifier.height(8.dp))
-        CardSwitchRow("Friends enabled", ui.enabled) { vm.setEnabled(it) }
-    }
-
-    // When off, this tab shows only the toggle above — nothing else is active.
-    if (!ui.enabled) return
-
-    Spacer(Modifier.height(12.dp))
-
-    // Incoming deep-link confirmation (adding from an eor:// link always needs explicit confirm).
-    ui.pendingLink?.let { parsed ->
-        SettingsSectionHeader("Friend request")
-        SettingsCard {
-            Text("Add ${parsed.displayName ?: parsed.deviceId.take(12) + "…"} as a friend?")
-            Spacer(Modifier.height(8.dp))
-            Row(horizontalArrangement = Arrangement.spacedBy(8.dp)) {
-                GradientFillButton(
-                    "Add",
-                    onClick = { vm.confirmPendingLink() },
-                    modifier = Modifier.weight(1f)
-                )
-                GradientOutlineButton(
-                    "Dismiss",
-                    onClick = { vm.dismissPendingLink() },
-                    modifier = Modifier.weight(1f)
-                )
-            }
-        }
-        Spacer(Modifier.height(8.dp))
-    }
-
-    // My code card
-    SettingsSectionHeader("My friend code")
-    SettingsCard {
-        var nameInput by rememberSaveable(ui.displayName) { mutableStateOf(ui.displayName) }
-        OutlinedTextField(
-            value = nameInput,
-            onValueChange = { nameInput = it },
-            label = { Text("Display name") },
-            singleLine = true,
-            modifier = Modifier.fillMaxWidth()
-        )
-        Spacer(Modifier.height(6.dp))
-        GradientOutlineButton(
-            "Save name",
-            onClick = { vm.saveDisplayName(nameInput) },
-            modifier = Modifier.fillMaxWidth()
-        )
-        Spacer(Modifier.height(10.dp))
-        when {
-            ui.engineStarting -> LoadingStatusRow(
-                "Starting the connection engine…",
-                MaterialTheme.colorScheme.onSurfaceVariant
-            )
-
-            ui.myShareLink != null -> {
-                Text(
-                    "Share this code with a friend. They add you, you add them back — then you're connected.",
-                    style = MaterialTheme.typography.labelSmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-                Spacer(Modifier.height(8.dp))
-                GradientFillButton(
-                    "Share my friend code",
-                    onClick = {
-                        val send = Intent(Intent.ACTION_SEND).apply {
-                            type = "text/plain"
-                            putExtra(Intent.EXTRA_TEXT, "Add me on eOr: ${ui.myShareLink}")
-                        }
-                        context.startActivity(Intent.createChooser(send, "Share friend code"))
-                    },
-                    modifier = Modifier.fillMaxWidth()
-                )
-            }
-
-            else -> Text("Bringing the engine up…", style = MaterialTheme.typography.labelSmall)
-        }
-    }
-
-    Spacer(Modifier.height(12.dp))
-
-    // Add a friend
-    SettingsSectionHeader("Add a friend")
-    SettingsCard {
-        OutlinedTextField(
-            value = codeInput,
-            onValueChange = { codeInput = it },
-            label = { Text("Paste a friend code or link") },
-            singleLine = true,
-            modifier = Modifier.fillMaxWidth()
-        )
-        Spacer(Modifier.height(6.dp))
-        GradientFillButton(
-            "Add friend",
-            onClick = { vm.addFriend(codeInput); codeInput = "" },
-            enabled = codeInput.isNotBlank(),
-            modifier = Modifier.fillMaxWidth()
-        )
-        ui.status?.let {
-            Spacer(Modifier.height(8.dp))
-            Text(
-                it,
-                style = MaterialTheme.typography.labelSmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant
-            )
-        }
-    }
-
-    Spacer(Modifier.height(12.dp))
-
-    // Add nearby (same Wi-Fi) — stop the beacon when this section leaves the screen.
-    DisposableEffect(Unit) { onDispose { vm.stopNearby() } }
-    SettingsSectionHeader("Add nearby")
-    SettingsCard {
-        Text(
-            "On the same Wi-Fi? Both of you open this, then tap each other to connect — no code needed.",
-            style = MaterialTheme.typography.labelSmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-        Spacer(Modifier.height(8.dp))
-        if (!ui.scanningNearby) {
-            GradientFillButton(
-                "Find nearby friends",
-                onClick = { vm.startNearby() },
-                modifier = Modifier.fillMaxWidth()
-            )
-        } else {
-            GradientOutlineButton(
-                "Stop searching",
-                onClick = { vm.stopNearby() },
-                modifier = Modifier.fillMaxWidth()
-            )
-            Spacer(Modifier.height(8.dp))
-            if (ui.nearby.isEmpty()) {
-                LoadingStatusRow(
-                    "Searching for nearby players…",
-                    MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            } else {
-                ui.nearby.forEachIndexed { i, peer ->
-                    if (i > 0) CardDivider()
-                    Row(
-                        Modifier
-                            .fillMaxWidth()
-                            .clip(RoundedCornerShape(8.dp))
-                            .clickable { vm.addNearby(peer) }
-                            .padding(vertical = 8.dp, horizontal = 4.dp),
-                        verticalAlignment = Alignment.CenterVertically,
-                        horizontalArrangement = Arrangement.SpaceBetween
-                    ) {
-                        Column(Modifier.weight(1f)) {
-                            Text(peer.displayName, style = MaterialTheme.typography.bodyMedium)
-                            // The name is unverified (anyone on the Wi-Fi can broadcast it); show the
-                            // start of the device id so the user can confirm who they're really adding.
-                            Text(
-                                "Unverified name · ID ${peer.deviceId.take(7)}…",
-                                style = MaterialTheme.typography.labelSmall,
-                                color = MaterialTheme.colorScheme.onSurfaceVariant
-                            )
-                        }
-                        Text(
-                            "Tap to add",
-                            style = MaterialTheme.typography.labelSmall,
-                            color = ElectricBlue
-                        )
-                    }
-                }
-            }
-        }
-    }
-
-    // Incoming requests
-    if (ui.incoming.isNotEmpty()) {
-        Spacer(Modifier.height(12.dp))
-        SettingsSectionHeader("Friend requests")
-        SettingsCard {
-            ui.incoming.forEachIndexed { i, f ->
-                if (i > 0) CardDivider()
-                Row(
-                    Modifier
-                        .fillMaxWidth()
-                        .padding(vertical = 6.dp),
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.SpaceBetween
-                ) {
-                    Text(
-                        f.displayName,
-                        style = MaterialTheme.typography.bodyMedium,
-                        modifier = Modifier.weight(1f)
-                    )
-                    Row(horizontalArrangement = Arrangement.spacedBy(6.dp)) {
-                        GradientFillButton("Accept", onClick = { vm.acceptRequest(f.deviceId) })
-                        GradientOutlineButton(
-                            "Decline",
-                            onClick = { vm.declineRequest(f.deviceId) })
-                    }
-                }
-            }
-        }
-    }
-
-    // My friends
-    Spacer(Modifier.height(12.dp))
-    SettingsSectionHeader("My friends")
-    SettingsCard {
-        val all = ui.active + ui.outgoing
-        if (all.isEmpty()) {
-            Text(
-                "No friends yet. Share your code to get started.",
-                style = MaterialTheme.typography.labelSmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant
-            )
-        } else {
-            all.forEachIndexed { i, f ->
-                if (i > 0) CardDivider()
-                Row(
-                    Modifier
-                        .fillMaxWidth()
-                        .padding(vertical = 6.dp),
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.SpaceBetween
-                ) {
-                    Column(Modifier.weight(1f)) {
-                        Text(f.displayName, style = MaterialTheme.typography.bodyMedium)
-                        Text(
-                            if (f.status == FriendStatus.PENDING_OUT) "Waiting for them to add you back" else "Connected",
-                            style = MaterialTheme.typography.labelSmall,
-                            color = MaterialTheme.colorScheme.onSurfaceVariant
-                        )
-                    }
-                    GradientOutlineButton("Remove", onClick = { vm.removeFriend(f.deviceId) })
-                }
-            }
-        }
-        Spacer(Modifier.height(8.dp))
-        GradientOutlineButton(
-            "Refresh",
-            onClick = { vm.refresh() },
-            modifier = Modifier.fillMaxWidth()
-        )
-    }
-}
